% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_qc_plot.R
\name{af.qc.plot}
\alias{af.qc.plot}
\title{Autofluorescence QC Plot}
\usage{
af.qc.plot(
  af.spectra,
  spectra,
  qc.table,
  af.color = "black",
  fluor.color = "blue",
  af.line.type = "solid",
  fluor.line.type = "dotted",
  linewidth = 1,
  plot.dir = "./figure_autofluorescence",
  filename = "autofluorescence_qc_report.pdf"
)
}
\arguments{
\item{af.spectra}{Matrix or dataframe containing autofluorescence spectral
signatures. This should be in format AFn x detectors. Row names will be used
as the names.}

\item{spectra}{Matrix or dataframe containing spectral data. This
should be in format fluorophores x detectors. Row names will be used as the
fluorophore names. Column names will be used as the detectors (channels).}

\item{qc.table}{Dataframe containing the pre-screened problematic AF spectra
and the fluorophores to which they are similar. Format: columes named \code{AF},
\code{Fluorophore} and \code{Similarity} containing, respectively, the names of the AF
spectra, the names of the similar fluorophores and the cosine similarity values.}

\item{af.color}{Color for the line representing the user's autofluorescence
spectrum from the unstained reference control. Default is \code{black}.}

\item{fluor.color}{Color for the line representing the similar fluorophore
spectrum. Default is \code{blue}.}

\item{af.line.type}{Line style for the line representing the user's
AF spectrum. Default is \code{solid}.}

\item{fluor.line.type}{Line style for the line representing thefluorophore
spectrum. Default is \code{dotted}.}

\item{linewidth}{Width of the line for the spectral traces. Default is \code{1}.}

\item{plot.dir}{Directory where the files will be saved.
Default is \code{./figure_autofluorescence}.}

\item{filename}{Name for the output PDF file. Default is
\code{autofluorescence_qc_report.pdf}.}
}
\value{
None. Plots are saved to a PDF in \code{plot.dir}.
}
\description{
This function plots a comparison between the user's fluorophore spectrum,
from \code{spectra}, and an identified autofluorescence spectrum that may represent
contamination. Quality control is performed using cosine similarity between the
two spectral profiles.
}
\seealso{
\itemize{
\item \code{\link[=qc.af.spectra]{qc.af.spectra()}}
\item \code{\link[=get.af.spectra]{get.af.spectra()}}
}
}
