% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_af_fluorophores.R
\name{assign.af.fluorophores}
\alias{assign.af.fluorophores}
\title{Assign AF Spectrum By Fluorophore Projection}
\usage{
assign.af.fluorophores(raw.data, spectra, af.spectra)
}
\arguments{
\item{raw.data}{Expression data from raw FCS files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns.}
}
\value{
Row indices for best-fitting AF spectra (from \code{af.spectra})
}
\description{
Projects the autofluorescence spectral variants into fluorophore unmixed
space to determine which best fits each cell (event). A fast approximation for
brute force sequential unmixing method in early versions of AutoSpectral.
Provides essentially identical results to minimization of fluorophore signal
(dist0 method), with minor differences being primarily due to the use of L2
(squared) error, rather than L1 (absolute) error. Substantially faster.
}
