% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_gates.R
\name{assign.gates}
\alias{assign.gates}
\title{Assign Gates}
\usage{
assign.gates(control.table, gating.system, gate, verbose = TRUE)
}
\arguments{
\item{control.table}{Dataframe or table of the control file, read in via
\code{define.flow.control()} and cleaned up by that function.}

\item{gating.system}{Character string selecting the automated gating system to
employ in defining the initial scatter gates for identifying cells in the FCS
files. Options are \code{landmarks} and \code{density}. The \code{density} option uses the
original gating system from AutoSpill, picking cell populations based on dense
regions on FSC/SSC. The default \code{landmarks} system picks out the cell regions
using the brightest events in the peak channel for the single-stained control(s).
This approach is generally more robust. A good way to use this is to utilize
\code{landmarks} in combination with specifying which single-stained control files
should be used for defining the gates. For example, abundant, bright cell
markers such as CD4 will reliably identify the lymphocyte region, as CD14 will
identify the monocyte region, etc. For more instructions, see the help pages
on GitHub.}

\item{gate}{Logical, default is \code{TRUE}, in which case, automated gating will
be performed. If \code{FALSE}, the FCS files will be imported without automatically
generated gates applied. That is, all data in the files will be used. This is
intended to allow the user to pre-gate the files in commercial software.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A modified \code{control.table}, complete with gate.name and gate.define
}
\description{
Assigns gate labels to samples based on whether they share characteristics,
such as being the same for \code{viability.gate}, \code{large.gate} or \code{control.type}
in the control file and whether the universal negative used is the same. Updates
the supplied control table and returns it with gate assignments.
}
\seealso{
\itemize{
\item \code{\link[=tune.gate]{tune.gate()}}
\item \code{\link[=define.flow.control]{define.flow.control()}}
}
}
