% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_variants_cosine.R
\name{assign.variants.cosine}
\alias{assign.variants.cosine}
\title{Assign Variant Spectrum By Fluorophore Projection}
\usage{
assign.variants.cosine(raw.data, spectra, fluorophore, variant.spectra)
}
\arguments{
\item{raw.data}{Expression data from raw FCS files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{fluorophore}{Name of the fluorophore to be tested for variation.}

\item{variant.spectra}{Spectral signatures of fluorophore variants, normalized
between 0 and 1, with fluorophores in rows and detectors in columns.}
}
\value{
Row indices for best-fitting variant spectra (from \code{variant.spectra})
}
\description{
Projects the autofluorescence spectral variants into fluorophore unmixed
space to determine which best fits each cell (event). A fast approximation for
brute force sequential unmixing method in early versions of AutoSpectral.
Provides essentially identical results to minimization of fluorophore signal
(dist0 method). Substantially faster. Use L1 (absolute value) minimization,
which works better than the standard L2 (squared error).
}
