% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSpectral-package.R
\docType{package}
\name{AutoSpectral-package}
\alias{AutoSpectral}
\alias{AutoSpectral-package}
\title{AutoSpectral: Tools for Unmxing Spectral Flow Cytometry Data}
\description{
To learn more about using the AutoSpectral package, please start by browsing the
vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrCytometer/AutoSpectral}
  \item Report bugs at \url{https://github.com/DrCytometer/AutoSpectral/issues}
}

}
\author{
\strong{Maintainer}: Oliver Burton \email{olivertburton@gmail.com} (\href{https://orcid.org/0000-0003-3884-7373}{ORCID})

Authors:
\itemize{
  \item Carlos P. Roca \email{carlosproca@gmail.com} (\href{https://orcid.org/0000-0003-0230-1926}{ORCID})
}

}
\keyword{internal}
