% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSpectral-package.R
\docType{package}
\name{AutoSpectral}
\alias{AutoSpectral-package}
\alias{AutoSpectral}
\title{AutoSpectral: Tools for Unmixing Spectral Flow Cytometry Data}
\description{
As both a refinement and advancement of the unmixing process of full spectrum cytometry data, \code{AutoSpectral} provides a suite of functions/tools that work in concert to provide the user with optimal unmixing results.

Use \code{AutoSpectral} to:
\itemize{
\item isolate and refine clean spectral signatures
\item mitigate autofluorescence contamination in complex controls
\item unmix fcs files using standard algorithms or the hallmark \code{AutoSpectral} approach
}
}
\section{\code{AutoSpectral} (v 1.5.7) Workflow}{

To maximize the success of \code{AutoSpectral}, the workflow is organized into the following logical step-wise processes:
\enumerate{
\item \link{create.control.file}: Generate a .csv file that contains descriptive information about your single stained controls
\itemize{
\item \href{https://drcytometer.github.io/AutoSpectral/articles/01_Full_AutoSpectral_Workflow.html#creating-the-control-file}{Control File} (see article)
}
\item \link{define.flow.control}: uses the result of \link{create.control.file} to convert FCS files and their associated metadata into an optimized data structure for downstream workflows
\itemize{
\item \href{https://drcytometer.github.io/AutoSpectral/articles/01_Full_AutoSpectral_Workflow.html#loading-the-data}{Flow Control} (see article)
}
\item ...
\item ...
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrCytometer/AutoSpectral}
  \item \url{https://drcytometer.github.io/AutoSpectral/}
  \item Report bugs at \url{https://github.com/DrCytometer/AutoSpectral/issues}
}

}
\author{
\strong{Maintainer}: Oliver Burton \email{olivertburton@gmail.com} (\href{https://orcid.org/0000-0003-3884-7373}{ORCID})

Authors:
\itemize{
  \item Adrian Liston \email{adrian.liston.com} (\href{https://orcid.org/0000-0002-6272-4085}{ORCID})
}

}
