% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity_index.R
\name{calculate.complexity.index}
\alias{calculate.complexity.index}
\title{Calculate Complexity Index}
\usage{
calculate.complexity.index(spectra)
}
\arguments{
\item{spectra}{A matrix of spectral signatures of fluorophores, normalized
between 0 and 1.
Fluorophores should be in rows and detectors (channels) in columns.
More than one spectrum is required to perform the calculation.}
}
\value{
A numeric value representing the complexity index (condition number)
for the spectral flow cytometry panel.
}
\description{
Calculates the complexity index (condition number) for a set of
fluorophores in a spectral panel.
}
