% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_condition_number.R
\name{calculate.condition.number}
\alias{calculate.condition.number}
\title{Calculate Condition Number}
\usage{
calculate.condition.number(spectra)
}
\arguments{
\item{spectra}{A matrix of spectral signatures of fluorophores, normalized
between 0 and 1.
Fluorophores should be in rows and detectors (channels) in columns.
More than one spectrum is required to perform the calculation.}
}
\value{
A numeric value representing the condition number for the spectral
flow cytometry panel.
}
\description{
Calculates the mixing matrix condition number ("Complexity IndexTM") for a
set of fluorophores in a spectral panel.
}
