% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hotspot_matrix.R
\name{calculate.hotspot.matrix}
\alias{calculate.hotspot.matrix}
\title{Calculate Hotspot Matrix}
\usage{
calculate.hotspot.matrix(spectra)
}
\arguments{
\item{spectra}{Fluorophore spectra as fluorophores x detectors}
}
\value{
The hotspot matrix
}
\description{
Calculates the hotspot (unmixing spread) matrix
}
