% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_channels.R
\name{check.channels}
\alias{check.channels}
\title{Check Channels}
\usage{
check.channels(spectral.channels, asp)
}
\arguments{
\item{spectral.channels}{Vector of initial spectral channel names.}

\item{asp}{The AutoSpectral parameter list. Generate using
\code{get.autospectral.param}}
}
\value{
Returns the vector of spectral channels re-organized in excitation-
emission order. That is, narrowest to longest excitation laser, wtih narrowest
to longest emission wavelength inside each laser group.
}
\description{
A helper function to reorganize the spectral channels in a nice order for
plotting. Puts them in excitation/emission order.
}
