% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gate_consistency.R
\name{check.consistency}
\alias{check.consistency}
\title{Check Gate Consistency}
\usage{
check.consistency(vec, var.name)
}
\arguments{
\item{vec}{Vector of a factor, such as \code{large.gate} settings.}

\item{var.name}{Name of the factor (variable) being checked for multiplicity}
}
\value{
The unique element in the factor \code{vec}, if all elements of \code{vec} are
identical.
}
\description{
A small helper function to check for non-unique input factors when setting
gate definitions.
}
\seealso{
\itemize{
\item \code{\link[=tune.gate]{tune.gate()}}
\item \code{\link[=define.gate.landmarks]{define.gate.landmarks()}}
\item \code{\link[=define.gate.density]{define.gate.density()}}
\item \code{\link[=do.gate]{do.gate()}}
\item \code{\link[=define.flow.control]{define.flow.control()}}
}
}
