% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_critical.r
\name{check.critical}
\alias{check.critical}
\title{Perform Critical Check}
\usage{
check.critical(condition, error.msg)
}
\arguments{
\item{condition}{Essential condition(s) evaluated as a logical TRUE/FALSE.}

\item{error.msg}{Message to be returned to the user if condition is FALSE.}
}
\value{
Returns error.msg and breaks if FALSE, no return if TRUE.
}
\description{
Checks condition, if not true prints error message and
stops execution.
}
