% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_fcs.R
\name{concatenateFCS}
\alias{concatenateFCS}
\title{Concatenate Multiple FCS Files}
\usage{
concatenateFCS(
  fcs.paths,
  output.name = "Concatenated.fcs",
  output.dir = "./concatenated_fcs"
)
}
\arguments{
\item{fcs.paths}{A character vector of full file paths to the input .fcs files.}

\item{output.name}{A character string for the name of the new .fcs file.
Default is \code{Concatenated.fcs}.}

\item{output.dir}{A character string specifying the directory to save the new
file. Default is \code{./concatenated_fcs}}
}
\value{
Path to the newly created FCS file.
}
\description{
Takes a string/list of input .fcs file paths, reads in the data, combines it
and writes a new .fcs file with all the data together.
}
