% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spillover_to_flowjo.R
\name{convert.spillover.to.flowjo}
\alias{convert.spillover.to.flowjo}
\title{Convert Spillover To FlowJo}
\usage{
convert.spillover.to.flowjo(
  spillover.file.name,
  flowjo.file.name,
  input.dir,
  output.dir
)
}
\arguments{
\item{spillover.file.name}{The file name of the spillover csv file to be imported.}

\item{flowjo.file.name}{The name given to the generated FlowJo compensation file.}

\item{input.dir}{Path to the spillover csv file.}

\item{output.dir}{Path to the folder where the compensation mtx file will be written.}
}
\value{
No return. Write the xml .mtx file.
}
\description{
A quick and dirty method for writing a FlowJo-compatible compensation matrix
file from any spillover csv file.
}
