% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity_plot.R
\name{cosine.similarity.plot}
\alias{cosine.similarity.plot}
\title{Cosine Similarity Plot}
\usage{
cosine.similarity.plot(
  spectra,
  filename = "autospectral_similarity_matrix",
  title = NULL,
  output.dir = "figure_similarity_heatmap",
  figure.width = 8,
  figure.height = 6,
  color.palette = "viridis",
  show.legend = TRUE,
  file.type = "jpg",
  save = TRUE
)
}
\arguments{
\item{spectra}{Data frame or matrix containing spectral data.}

\item{filename}{Character string for the output file. Default is
\code{autospectral_similarity_matrix}.}

\item{title}{Optional prefix for the plot filename. Default is \code{NULL}}

\item{output.dir}{File path where the plot will be created. Default is
\code{figure_similarity_heatmap}. The directory will be created if it does not
already exist.}

\item{figure.width}{Numeric. Width of output plot. Default is \code{8}.}

\item{figure.height}{Numeric. Height of output plot. Default is \code{6}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{viridis}. Options
are the viridis color options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis},
\code{cividis}, \code{rocket}, \code{mako} and \code{turbo}.}

\item{show.legend}{Logical. If \code{TRUE}, figure legend will be included.}

\item{file.type}{Character string specifying the output file format. One of
\code{"jpg"} (default), \code{"jpeg"}, \code{"tiff"}, \code{"png"}, or \code{"pdf"}.}

\item{save}{Logical, if \code{TRUE}, saves a file to the \code{output.dir}.
Otherwise, the plot will simply be created in the Viewer.}
}
\value{
Saves the heatmap plot as an image file in the similarity directory.
}
\description{
This function plots the matrix of cosine similarities (AKA "Similarity Matrix")
as a heatmap and saves it as a JPEG file. It also calculates and displays the
mixing matrix condition number (AKA "Complexity Index") of the matrix.
}
