% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine.similarity}
\alias{cosine.similarity}
\title{Calculate Cosine Similarity}
\usage{
cosine.similarity(spectra)
}
\arguments{
\item{spectra}{The spectral matrix (or dataframe), represented as
fluorophores in rows and detectors in columns.}
}
\value{
The cosine similarity matrix in fluorophores x fluorophores.
}
\description{
Calculates the cosine similarity between columns of the input matrix (spectra).
}
