% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.r
\name{create.directory}
\alias{create.directory}
\title{Create Directory}
\usage{
create.directory(asp)
}
\arguments{
\item{asp}{The AutoSpectral parameter list defined
using \code{get.autospectral.param}}
}
\value{
No returns. Creates directories in the current working directory.
}
\description{
Creates figure and table directories.
}
