% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_gate_density.R
\name{define.gate.density}
\alias{define.gate.density}
\title{Define Gate by Density}
\usage{
define.gate.density(
  control.file,
  control.dir,
  asp,
  gating.params = NULL,
  n.cells = NULL,
  grid.n = NULL,
  bandwidth.factor = NULL,
  target.pop = NULL,
  neighbors = NULL,
  fsc.channel = NULL,
  ssc.channel = NULL,
  fsc.lims = NULL,
  ssc.lims = NULL,
  fsc.search.min = NULL,
  fsc.search.max = NULL,
  ssc.search.min = NULL,
  ssc.search.max = NULL,
  output.dir = "./figure_gate",
  gate.name = "cell_gate",
  filename = "density_gate_definition_",
  color.palette = "plasma",
  boundary.color = "black",
  points.to.plot = 1e+05,
  width = 5,
  height = 5
)
}
\arguments{
\item{control.file}{File path and name for the CSV file defining the single-
color control file names, fluorophores they represent, marker names, peak
channels, and gating requirements.}

\item{control.dir}{File path to the single-stained control FCS files.}

\item{asp}{The AutoSpectral parameter list defined using
\code{get.autospectral.param}.}

\item{gating.params}{Previously saved gating parameters. Load in the .rds file
using \code{readRDS} and pass the result here if you wish to replicate a previous
run. This is essentially just an updated version of \code{asp} containing any
modifications based on information passed as arguments to \code{define.gate.landmarks}.}

\item{n.cells}{The number of cells to use for defining the gate boundary. This
many cells will be selected from the specified FCS files. The default, \code{NULL},
will revert to \code{asp$gate.downsample.n}. Around 1e5 is recommended.}

\item{grid.n}{Numeric, default \code{100}. The binning grid for the kernel density
estimation. If \code{n.cells} is very low, you may wish to lower this number to
compress the search space.}

\item{bandwidth.factor}{Numeric, default \code{NULL} will revert to
\code{asp$gate.bound.density.bw.factor} (usually 1). A multiplier for the
bandwidth for the kernel density estimation. Larger numbers will smooth the
density, reducing discrimination between peaks in the density (such as between
live and dead cells).}

\item{target.pop}{Numeric 0-n, default \code{NULL} will revert to
\code{asp$gate.bound.density.max.target} (usually 1). In the output plots, the
numbers correspond to density peaks. Whichever number is set here will select
the population 1 less than that number as the target for defining the gate
region.}

\item{neighbors}{Number of neighboring events to consider when finding local
density maxima in the kernel density estimation. Default \code{NULL} will revert
to \code{asp$gate.bound.density.neigh.size}, which is usually 3.}

\item{fsc.channel}{Channel to use for Forward Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[1]}, which is appropriate for your
machine.}

\item{ssc.channel}{Channel to use for Side Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[2]}, which is appropriate for your
machine. For machines with multiple side scatter measurements, you can change
this.}

\item{fsc.lims}{Numeric vector. Limits for searching and plotting the FSC. The
default \code{NULL} uses c\code{c(asp$scatter.data.min.x, asp$scatter.data.max.x)}.}

\item{ssc.lims}{Numeric vector. Limits for searching and plotting the SSC. The
default \code{NULL} uses c\code{c(asp$scatter.data.min.y, asp$scatter.data.max.y)}.}

\item{fsc.search.min}{The minimum bound for the search space for dense event
populations for defining the gate on the x-axis. The default \code{NULL} reverts to
\code{asp$default.gate.param$region.factor.x.low}, which is usually 0.05 or 5\% of
the FSC axis.}

\item{fsc.search.max}{The maximum bound for the search space for dense event
populations for defining the gate on the x-axis. The default \code{NULL} reverts to
\code{asp$default.gate.param$region.factor.x.high}, which is usually 0.8 or 80\% of
the FSC axis.}

\item{ssc.search.min}{The minimum bound for the search space for dense event
populations for defining the gate on the y-axis. The default \code{NULL} reverts to
\code{asp$default.gate.param$region.factor.y.low}, which is usually 0.05 or 5\% of
the SSC axis.}

\item{ssc.search.max}{The maximum bound for the search space for dense event
populations for defining the gate on the y-axis. The default \code{NULL} reverts to
\code{asp$default.gate.param$region.factor.y.high}, which is usually 0.8 or 80\% of
the SSC axis.#'}

\item{output.dir}{File path where you want to save the results. Default is
\code{./figure_gate}.}

\item{gate.name}{Character, name for the gate. Useful for distinguishing gates
when you have multiple types. Default is \code{cell_gate}.}

\item{filename}{Character, name for the output files. Default is \code{gate_definition}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{plasma}. Use \code{rainbow}
to be similar to FlowJo or SpectroFlo. Other options are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{boundary.color}{Color for the gate boundary line on the plot. Default
is \code{black}.}

\item{points.to.plot}{Numeric, default \code{1e5}. Maximum number of points to show
on the plot. More points will take longer, but you really shouldn't have even
close to this number when defining the gate with landmarks.}

\item{width}{Numeric, default \code{5}. Width of the saved plot.}

\item{height}{Numeric, default \code{5}. Height of the saved plot.}
}
\value{
A set of points describing the gate boundary.
}
\description{
Performs gating on scatter parameters based on a density search.

The gating proceeds in three steps:
\itemize{
\item Defines bounds by data trimming
\item Defines a region around the target maximum found within the bounds
\item Defines a gate around the target maximum, only within that region
}

The method uses numerical search of maxima over estimated densities and
Voronoi tessellations to improve density estimation around maxima.
}
\references{
Roca, Carlos P et al. "AutoSpill is a principled framework that
simplifies the analysis of multichromatic flow cytometry data" \emph{Nature
Communications} 12 (2890) 2021.
}
\seealso{
\itemize{
\item \code{\link[=tune.gate]{tune.gate()}}
\item \code{\link[=define.gate.landmarks]{define.gate.landmarks()}}
\item \code{\link[=do.gate]{do.gate()}}
\item \code{\link[=gate.define.plot]{gate.define.plot()}}
}
}
