% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_gate_landmarks.R
\name{define.gate.landmarks}
\alias{define.gate.landmarks}
\title{Define Gate by Landmarks}
\usage{
define.gate.landmarks(
  control.file,
  control.dir,
  asp,
  gating.params = NULL,
  n.cells = 2000,
  percentile = 70,
  grid.n = 100,
  bandwidth.factor = 1,
  fsc.channel = NULL,
  ssc.channel = NULL,
  fsc.lims = NULL,
  ssc.lims = NULL,
  output.dir = "./figure_gate",
  gate.name = "cell_gate",
  filename = "landmark_gate_definition_",
  color.palette = "plasma",
  boundary.color = "black",
  points.to.plot = 1e+05,
  width = 5,
  height = 5
)
}
\arguments{
\item{control.file}{File path and name for the CSV file defining the single-
color control file names, fluorophores they represent, marker names, peak
channels, and gating requirements.}

\item{control.dir}{File path to the single-stained control FCS files.}

\item{asp}{The AutoSpectral parameter list defined using
\code{get.autospectral.param}.}

\item{gating.params}{Previously saved gating parameters. Load in the .rds file
using \code{readRDS} and pass the result here if you wish to replicate a previous
run. This is essentially just an updated version of \code{asp} containing any
modifications based on information passed as arguments to \code{define.gate.landmarks}.}

\item{n.cells}{The number of cells to use for defining the gate boundary. This
many cells will be selected from the peak channel (brightest first) in the
single-color controls. For example, if you set \code{200} and marked files such as
\code{CD3-PE.fcs} and \code{CD19-FITC.fcs} as \code{gate.define=TRUE} in the control file,
the brightest 200 events in the YG1 channel from the CD3-PE file and the
brightest 200 events in the B1 channel for the CD19-FITC file would be used
to define the gate.}

\item{percentile}{Numeric 1 - 100, default \code{70}. The percentile cutoff for
density in the scatter to use for defining the gate. For example, a value of
\code{50} would take the 50\% of cells closest to the density peak, more or less.
Smaller numbers will define a tighter gate.}

\item{grid.n}{Numeric, default \code{100}. The binning grid for the kernel density
estimation. If \code{n.cells} is very low, you may wish to lower this number to
compress the search space.}

\item{bandwidth.factor}{Numeric, default \code{1}. A multiplier for the bandwidth
for the kernel density estimation. Larger numbers will smooth the density,
reducing discrimination between peaks in the density (such as between live
and dead cells).}

\item{fsc.channel}{Channel to use for Forward Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[1]}, which is appropriate for your
machine.}

\item{ssc.channel}{Channel to use for Side Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[2]}, which is appropriate for your
machine. For machines with multiple side scatter measurements, you can change
this.}

\item{fsc.lims}{Numeric vector. Limits for searching and plotting the FSC. The
default \code{NULL} uses c\code{c(asp$scatter.data.min.x, asp$scatter.data.max.x)}.}

\item{ssc.lims}{Numeric vector. Limits for searching and plotting the SSC. The
default \code{NULL} uses c\code{c(asp$scatter.data.min.y, asp$scatter.data.max.y)}.}

\item{output.dir}{File path where you want to save the results. Default is
\code{./figure_gate}.}

\item{gate.name}{Character, name for the gate. Useful for distinguishing gates
when you have multiple types. Default is \code{cell_gate}.}

\item{filename}{Character, name for the output files. Default is \code{gate_definition}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{plasma}. Use \code{rainbow}
to be similar to FlowJo or SpectroFlo. Other options are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{boundary.color}{Color for the gate boundary line on the plot. Default
is \code{black}.}

\item{points.to.plot}{Numeric, default \code{1e5}. Maximum number of points to show
on the plot. More points will take longer, but you really shouldn't have even
close to this number when defining the gate with landmarks.}

\item{width}{Numeric, default \code{5}. Width of the saved plot.}

\item{height}{Numeric, default \code{5}. Height of the saved plot.}
}
\value{
A set of points describing the gate boundary.
}
\description{
Performs gating on scatter parameters using events selected on the basis of
high fluorescence intensity in the expected peak channel in the single-stained
control samples. This approach allows for identification of cells fairly
reliably by "landmarks", especially when using well-defined, abundant
populations such a T cells (e.g., CD3, CD4, CD8 single-stained samples),
monocytes (CD14-stained sample) or neutrophils (CD66, CD33).
}
\references{
Laniewski, Nathan. \emph{flowstate}.
\url{https://github.com/nlaniewski/flowstate}
}
\seealso{
\itemize{
\item \code{\link[=tune.gate]{tune.gate()}}
\item \code{\link[=define.gate.density]{define.gate.density()}}
\item \code{\link[=gate.define.plot]{gate.define.plot()}}
}
}
