% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate_af.r
\name{do.gate.af}
\alias{do.gate.af}
\title{Perform Gating on Autofluorescence Parameters}
\usage{
do.gate.af(gate.data, samp, asp)
}
\arguments{
\item{gate.data}{A data frame containing the gate data.}

\item{samp}{A sample identifier.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.
A list of essential parameters.}
}
\value{
A vector with the indexes of events inside the initial gate.
}
\description{
Returns a vector with the indexes of events inside the initial
    gate on autofluorescence parameters.
}
\details{
Perform Gating on Autofluorescence Parameters

This function returns a vector with the indexes of events inside the initial gate on
    autofluorescence parameters. It proceeds through several steps to define the gate boundaries
    and identify density maxima using numerical search and Voronoi tessellations.
}
