% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate_test.R
\name{do.gate.test}
\alias{do.gate.test}
\title{Do Gate}
\usage{
do.gate.test(
  gate.data,
  viability.gate,
  large.gate,
  samp,
  scatter.and.channel.label,
  control.type,
  asp,
  color.palette = "plasma",
  max.points = 50000,
  gate.color = "darkgoldenrod1"
)
}
\arguments{
\item{gate.data}{A data frame containing the gate data.}

\item{viability.gate}{A logical vector indicating the viability gate.}

\item{large.gate}{A logical vector indicating the large gate.}

\item{samp}{A sample identifier.}

\item{scatter.and.channel.label}{A label for scatter and channel.}

\item{control.type}{The type of control used, either "beads" or "cells".}

\item{asp}{The AutoSpectral parameter list, prepared using
\code{get.autospectral.param}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{plasma}. Use \code{rainbow}
to be similar to FlowJo or SpectroFlo. Other options are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{max.points}{Number of points to plot (speeds up plotting). Default is
\code{5e4}.}

\item{gate.color}{Color to plot the gate boundary line, default is \code{darkgoldenrod1}.}
}
\value{
A set of points describing the gate boundary.
}
\description{
Performs gating on scatter parameters and returns a vector with the points
describing the gate boundary.

The gating proceeds in three steps:
\itemize{
\item Defines bounds by data trimming
\item Defines a region around the target maximum found within the bounds
\item Defines a gate around the target maximum, only within that region
}

The method uses numerical search of maxima over estimated densities and
Voronoi tessellations to improve density estimation around maxima.
}
