% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_peacoqc.R
\name{do.peacoQC}
\alias{do.peacoQC}
\title{Perform Quality Control on Flow Cytometry Data using PeacoQC}
\usage{
do.peacoQC(
  dirty.expr,
  sample.name,
  spectral.channel,
  biexp.transform,
  transform.inv,
  output.dir,
  time.param,
  all.channels,
  method = asp$peacoqc.method,
  figures = asp$figures
)
}
\arguments{
\item{dirty.expr}{A matrix containing the raw expression data, pre-cleaning.}

\item{sample.name}{The name of the sample.}

\item{spectral.channel}{The spectral channels to be used.}

\item{biexp.transform}{The biexponential transformation function.}

\item{transform.inv}{The inverse transformation function.}

\item{output.dir}{The directory to save output files.}

\item{time.param}{The time channel parameter.}

\item{all.channels}{A vector of all channels to be included in the final
cleaned data.}

\item{method}{The PeacoQC method to use. Inherited from \code{PeacoQC}. Options
are \code{all}, \code{MAD} or \code{IT}. Default is \code{asp$peacoqc.method}, which by default
is \code{MAD}.}

\item{figures}{Logical. Controls whether PeacoQC plots are created. Default
is \code{asp$figures}, for which the default is \code{TRUE}.}
}
\value{
A matrix with the cleaned expression data.
}
\description{
This function performs quality control on flow cytometry data using the
PeacoQC method. It transforms the data, removes margins, and identifies
good cells. The function also optionally plots the results and saves
the cleaned data.
}
