% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{.cosine.sim.rows}
\alias{.cosine.sim.rows}
\title{dot Cosine Similarity Rows}
\usage{
.cosine.sim.rows(mat, ref.vec)
}
\arguments{
\item{spectra}{The matrix (or dataframe), represented as events in rows and
detectors in columns.}
}
\value{
Returns a numeric vector of length nrow(mat).
}
\description{
Cosine similarity of each row of mat against a single reference vector.
}
