% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_issue.R
\name{.new_issue}
\alias{.new_issue}
\title{New Issue}
\usage{
.new_issue(severity, rule, filename = NA, column = NA, message)
}
\arguments{
\item{severity}{Character, severity of the issue, e.g., "error", "warning".}

\item{rule}{Character, rule being broken.}

\item{filename}{Character, name of the FCS file linked to the issue.}

\item{column}{Character, column of the control file linked to the issue.}

\item{message}{Character, output message to guide the user in fixing the error.}
}
\value{
A dataframe of issues
}
\description{
Helper function to log issues in control file
}
