% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_spread.R
\name{estimate.spread}
\alias{estimate.spread}
\title{Estimate Spread}
\usage{
estimate.spread(spectra)
}
\arguments{
\item{spectra}{The spectral matrix, dimensions fluorophores x detectors.}
}
\value{
The estimated spread matrix
}
\description{
A crude approach to predicting spillover spread in the unmixed space based
on the spectral matrix.
}
