% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_af_spline.R
\name{fit.af.spline}
\alias{fit.af.spline}
\title{Fit Spline to Autofluorescence Data}
\usage{
fit.af.spline(af.cells, non.af.cells, asp)
}
\arguments{
\item{af.cells}{A matrix containing the autofluorescence data.}

\item{non.af.cells}{A matrix containing the low autofluorescence data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
The boundaries of the events within the specified number of standard
deviations from the spline.
}
\description{
This function fits a spline to autofluorescence data, removing extreme events
and defining bounds equally far from zero. It uses robust linear modeling
and identifies events within a specified number of standard deviations from
the spline.
}
