% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_robust_linear_model.r
\name{fit.robust.linear.model}
\alias{fit.robust.linear.model}
\title{Fit Robust Linear Model}
\usage{
fit.robust.linear.model(
  x.data,
  y.data,
  x.name,
  y.name,
  max.iter = 100,
  fix.unmix = FALSE
)
}
\arguments{
\item{x.data}{A vector containing the predictor variable data.}

\item{y.data}{A vector containing the response variable data.}

\item{x.name}{The name of the predictor variable.}

\item{y.name}{The name of the response variable.}

\item{max.iter}{Numeric. Maximum number of iterations for the robust linear
model. Default is \code{100}.}

\item{fix.unmix}{Logical, default is \code{FALSE}. If \code{TRUE}, sets coefficient to
zero in case of failed convergence. Used for \code{fix.my.unmix.}}
}
\value{
A vector with the intercept and coefficient.
}
\description{
Returns a matrix by rows, with the intercept and coefficient of a robust
linear model fitted to the input data.
Reverts to a standard linear model in case of no convergence.
}
