% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_af_identify_plot.R
\name{gate.af.identify.plot}
\alias{gate.af.identify.plot}
\title{Plot Autofluorescence Identification Gate}
\usage{
gate.af.identify.plot(gate.data, samp, gate.region, gate.bound.density, asp)
}
\arguments{
\item{gate.data}{Matrix containing autofluorescence data points.}

\item{samp}{Sample identifier.}

\item{gate.region}{Dataframe containing region boundary information.}

\item{gate.bound.density}{Density (e.g., from MASS:kde2d) information}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the sample being used to identify intrusive
autofluorescence in the single-stained controls. The input data are expected
to be PCA projections of the unstained sample with an accompanying region to
identify the low-autofluorescence cell region.
}
