% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gate_af.r
\name{gate.af.plot}
\alias{gate.af.plot}
\title{Plot Autofluorescence Gate}
\usage{
gate.af.plot(samp, gate.data, gate.bound, gate.region, gate.population, asp)
}
\arguments{
\item{samp}{Sample identifier.}

\item{gate.data}{Data frame containing gate data points.}

\item{gate.bound}{List containing gate boundary information.}

\item{gate.region}{List containing gate region information.}

\item{gate.population}{List containing gate population information.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the AF gate, including intermediate steps,
    using ggplot2 and other necessary packages.
}
