% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gate_af_sample.R
\name{gate.af.sample.plot}
\alias{gate.af.sample.plot}
\title{Plot Gate Autofluorescence Sample}
\usage{
gate.af.sample.plot(samp, af.data, af.boundary.lower, af.boundary.upper, asp)
}
\arguments{
\item{samp}{Sample identifier.}

\item{af.data}{Matrix containing autofluorescence data points.}

\item{af.boundary.lower}{Matrix containing lower boundary information.}

\item{af.boundary.upper}{Matrix containing upper boundary information.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the gate autofluorescence sample,
    including upper and lower boundaries, using ggplot2 and other necessary
    packages.
}
