% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_define_plot.r
\name{gate.define.plot}
\alias{gate.define.plot}
\title{Gate Definition Plot}
\usage{
gate.define.plot(
  samp,
  gate.data,
  gate.marker,
  gate.bound,
  gate.region,
  gate.population,
  scatter.and.channel.label,
  asp,
  color.palette = "rainbow",
  max.points = 1e+05
)
}
\arguments{
\item{samp}{Sample identifier.}

\item{gate.data}{Matrix containing gate data points.}

\item{gate.marker}{Vector containing gate marker names.}

\item{gate.bound}{List containing gate boundary information.}

\item{gate.region}{List containing gate region information.}

\item{gate.population}{List containing gate population information.}

\item{scatter.and.channel.label}{Named vector mapping scatter and
channel labels.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{rainbow}, which will
be similar to FlowJo or SpectroFlo. Other pptions are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{max.points}{Number of points to plot (speeds up plotting). Default is
\code{1e5}.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the gate during the definition step,
including boundaries and regions.
}
