% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_scatter_match.R
\name{gate.scatter.match}
\alias{gate.scatter.match}
\title{Gate Matching Scatter}
\usage{
gate.scatter.match(scatter.data, percentile = 0.5, bw.factor = 1, grid.n = 10)
}
\arguments{
\item{scatter.data}{Matrix of forward and side scatter data}

\item{percentile}{Numeric 0-1. Percentile of density to use for defining the
contour for the gate boundary. Default is \code{0.5}.}

\item{bw.factor}{Numeric, scalar bandwidth multiplier. Larger numbers smooth
the density, expanding the gate outwards. Default is \code{1}.}

\item{grid.n}{Numeric, size of the grid for defining the kernel density
estimate. Default is \code{10}.}
}
\value{
A trimesh object gate boundary
}
\description{
Calculates the density of the input scatter coordinates, cuts at the defined
density percentile and calculates a convex hull boundary around the points
within the dense region. Permits selection of events with similar forward
and side scatter profiles in a paired sample, e.g., an unstained sample.
}
