% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_a5se.r
\name{get.autospectral.param.a5se}
\alias{get.autospectral.param.a5se}
\title{Get AutoSpectral Parameters for the FACSymphony A5SE Cytometer}
\usage{
get.autospectral.param.a5se(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the A5SE cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
