% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_aurora.r
\name{get.autospectral.param.aurora}
\alias{get.autospectral.param.aurora}
\title{Get Autospectral Parameters for Aurora Cytometer}
\usage{
get.autospectral.param.aurora(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the Aurora cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
\details{
Get Autospectral Parameters for Aurora Cytometer

This function returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
