% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_auroraNL.r
\name{get.autospectral.param.auroraNL}
\alias{get.autospectral.param.auroraNL}
\title{Get Autospectral Parameters for the Aurora Northern Lights Cytometer}
\usage{
get.autospectral.param.auroraNL(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the Aurora cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
