% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_id7000.R
\name{get.autospectral.param.id7000}
\alias{get.autospectral.param.id7000}
\title{Get Autospectral Parameters for ID7000 Cytometer}
\usage{
get.autospectral.param.id7000(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the ID7000 cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral for the ID7000, without creating any figures or tables.
}
