% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_minimal.r
\name{get.autospectral.param.minimal}
\alias{get.autospectral.param.minimal}
\title{Get Minimal Autospectral Parameters}
\usage{
get.autospectral.param.minimal()
}
\value{
A list of minimal autospectral parameters.
}
\description{
Returns a minimal set of autospectral parameters.
}
\details{
Get Minimal Autospectral Parameters

This function returns a minimal set of autospectral parameters.
}
