% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_mosaic.r
\name{get.autospectral.param.mosaic}
\alias{get.autospectral.param.mosaic}
\title{Get Autospectral Parameters for Mosaic Cytometer}
\usage{
get.autospectral.param.mosaic(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the Mosaic cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
