% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_opteon.r
\name{get.autospectral.param.opteon}
\alias{get.autospectral.param.opteon}
\title{Get Autospectral Parameters for Opteon Cytometer}
\usage{
get.autospectral.param.opteon(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the Opteon cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral for the Opteon, without creating any figures or tables.
}
