% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_s8.r
\name{get.autospectral.param.s8}
\alias{get.autospectral.param.s8}
\title{Get Autospectral Parameters for DiscoverS8 Cytometer}
\usage{
get.autospectral.param.s8(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial autospectral parameters.}
}
\value{
A list of autospectral parameters specific to the DiscoverS8 cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
    autospectral, without creating any figures or tables.
}
\details{
Get Autospectral Parameters for DiscoverS8 Cytometer

This function returns parameters for running a calculation of unmixing with
    autospectral, without creating any figures or tables.
}
