% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_xenith.r
\name{get.autospectral.param.xenith}
\alias{get.autospectral.param.xenith}
\title{Get Autospectral Parameters for Xenith Cytometer}
\usage{
get.autospectral.param.xenith(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the Xenith cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
