% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density_palette.r
\name{get.density.palette}
\alias{get.density.palette}
\title{Get Density Color Palette}
\usage{
get.density.palette(dens, asp)
}
\arguments{
\item{dens}{A numeric vector representing the density distribution.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A vector of colors representing the density palette.
}
\description{
Returns a good color palette for a given density distribution.
}
\details{
Get Density Color Palette

This function returns a good color palette for a given density distribution.
}
