% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gated_flow_expression_data.R
\name{get.gated.flow.expression.data}
\alias{get.gated.flow.expression.data}
\title{Get Gated Flow Expression Data}
\usage{
get.gated.flow.expression.data(
  samp,
  file.name,
  control.dir,
  scatter.and.spectral.channel,
  spectral.channel,
  set.resolution,
  flow.gate,
  gate.list,
  scatter.param,
  scatter.and.channel.label,
  asp,
  apply.gate
)
}
\arguments{
\item{samp}{The sample identifier.}

\item{file.name}{A vector of file names for the samples.}

\item{control.dir}{The directory containing the control files.}

\item{scatter.and.spectral.channel}{A vector of scatter and spectral channels.}

\item{spectral.channel}{A vector of spectral channels.}

\item{set.resolution}{The resolution limit for the spectral channels.}

\item{flow.gate}{A list of flow gates for the samples.}

\item{gate.list}{A list of gating boundaries.}

\item{scatter.param}{A vector of scatter parameters.}

\item{scatter.and.channel.label}{A label for scatter and channel.}

\item{asp}{The AutoSpectral parameter list.}

\item{apply.gate}{Logical, whether to apply a scatter gate (supplied to \code{gate.list}).}
}
\value{
A matrix with the gated expression data.
}
\description{
Retrieves gated flow cytometry expression data for specified
samples, removing out-of-range events and applying gating boundaries.
}
