% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_events.R
\name{get.top.events}
\alias{get.top.events}
\title{Get Top (Brightest) Events}
\usage{
get.top.events(fcs.path, channel, n = 2000, scatter.param)
}
\arguments{
\item{fcs.path}{File path and name for the FCS file to be read.}

\item{channel}{Channel (detector) name for the selection of the brightest events.}

\item{n}{Number of brightest events to retrieve, default is \code{2000}.}

\item{scatter.param}{The names of the scatter channels to define which data
are returned.}
}
\value{
Returns a matrix of scatter (FSC, SSC) data
}
\description{
Retrieves the brightest \code{n} events in the specified FCS file in the specified
\code{channel}.
}
