% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ungated_flow_expression_data.R
\name{get.ungated.flow.expression.data}
\alias{get.ungated.flow.expression.data}
\title{Get Ungated Flow Expression Data}
\usage{
get.ungated.flow.expression.data(
  samp,
  file.name,
  control.dir,
  scatter.and.spectral.channel,
  spectral.channel,
  set.resolution
)
}
\arguments{
\item{samp}{The sample identifier.}

\item{file.name}{A vector of file names for the samples.}

\item{control.dir}{The directory containing the control files.}

\item{scatter.and.spectral.channel}{A vector of scatter and spectral channels.}

\item{spectral.channel}{A vector of spectral channels.}

\item{set.resolution}{The resolution limit for the spectral channels.}
}
\value{
A matrix with the flow expression data.
}
\description{
Retrieves flow cytometry expression data for specified samples, without
gating, removing out-of-range events.
}
