% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_universal_negative_test.R
\name{get.universal.negative.test}
\alias{get.universal.negative.test}
\title{Get Universal Negative Control}
\usage{
get.universal.negative.test(
  clean.expr.data,
  samp,
  universal.negatives,
  scatter.param,
  peak.channels,
  downsample,
  negative.n,
  positive.n,
  spectral.channel,
  asp,
  control.type,
  scatter.match = TRUE,
  k.neighbors = 3L,
  intermediate.figures = FALSE,
  main.figures = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{clean.expr.data}{List containing cleaned expression data.}

\item{samp}{Sample identifier.}

\item{universal.negatives}{Named vector mapping samples to their universal
negatives.}

\item{scatter.param}{Vector of scatter parameters.}

\item{peak.channels}{Named vector mapping samples to their peak channels.}

\item{downsample}{Logical indicating whether to downsample the data.}

\item{negative.n}{Number of negative events to select.}

\item{positive.n}{Number of positive events to select.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{control.type}{Named vector mapping samples to their control types.}

\item{scatter.match}{Logical indicating whether to perform scatter matching.
Default is \code{TRUE}.}

\item{k.neighbors}{Numeric, number of scatter-matched unstained events to
pair with every positive event for background determination. Default is \code{3}.}

\item{intermediate.figures}{Logical, if \code{TRUE} returns additional figures to
show the inner workings of the cleaning, including definition of low-AF cell
gates on the PCA-unmixed unstained and spectral ribbon plots of the AF
exclusion from the unstained. Default is \code{FALSE} to speed up processing.}

\item{main.figures}{Logical, if \code{TRUE} creates the main figures to show the
impact of intrusive autofluorescent event removal and scatter-matching for
the negatives.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A data frame containing the selected positive and scatter-matched
negative events.
}
\description{
This function identifies and processes the universal negative
control for a given sample, including scatter matching and plotting.
}
