% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unmixing_error.R
\name{get.unmixing.error}
\alias{get.unmixing.error}
\title{Calculate Unmixing Error}
\usage{
get.unmixing.error(
  expr.data.unmix,
  fluorophores,
  scale.untransformed,
  transform.inv,
  flow.event.sample,
  asp
)
}
\arguments{
\item{expr.data.unmix}{Data frame containing unmixed expression data.}

\item{fluorophores}{Vector of fluorophore names.}

\item{scale.untransformed}{Logical indicating whether to scale untransformed
data.}

\item{transform.inv}{Function to apply the inverse transformation.}

\item{flow.event.sample}{Vector indicating the sample for each flow event.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A list containing the unmixing correction matrices for intercept,
coefficient, slope, and skewness.
}
\description{
This function calculates the unmixing error for each
fluorophore, including intercept, coefficient, slope, and skewness, using
robust linear modeling.
}
