% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gating_error.R
\name{handle.gating.error}
\alias{handle.gating.error}
\title{Handle and Plot Gating Failures}
\usage{
handle.gating.error(
  e,
  gate.id,
  files.to.gate,
  scatter.coords,
  samp,
  viability.gate,
  control.type,
  flow.scatter.and.channel.label,
  asp
)
}
\arguments{
\item{e}{The error object from tryCatch}

\item{gate.id}{The ID of the gate that failed}

\item{files.to.gate}{Character vector of filenames used}

\item{scatter.coords}{The data matrix used for gating}

\item{samp}{String identifier for the sample/gate type}

\item{viability.gate}{Logical, is this a viability gate}

\item{control.type}{"beads" or "cells"}

\item{flow.scatter.and.channel.label}{Mapping vector for axis labels}

\item{asp}{The parameter list}
}
\value{
None. Stops flow.
}
\description{
A helper function that dictates what occurs when a gating error is detected.
}
