% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_af_artefacts.R
\name{identify.af.artefacts}
\alias{identify.af.artefacts}
\title{Identify Autofluorescence Artefacts}
\usage{
\method{identify}{af.artefacts}(clean.expr, universal.neg, spectral.channel, asp)
}
\arguments{
\item{clean.expr}{List containing cleaned expression data.}

\item{universal.neg}{Name of the universal negative control.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A list containing the autofluorescence components and gate boundaries.
}
\description{
This function identifies spikes of autofluorescence based on a
    negative control, defines gates for removal, plots removal gates, plots
    spectra of cells with and without removal, and returns gate boundaries
    and the components needed to transform single-stained controls.
}
