% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_unmix.R
\name{optimize.unmix}
\alias{optimize.unmix}
\title{Optimize Spectral Unmixing}
\usage{
optimize.unmix(
  raw.data,
  unmixed,
  unmix.fun,
  combined.spectra,
  weights,
  pos.thresholds,
  af.idx,
  af.spectra,
  optimize.fluors,
  variants,
  delta.list,
  delta.norms,
  fluorophores,
  af.idx.in.spectra,
  asp,
  k = 10L,
  cell.weighting = FALSE,
  cell.weight.regularize = FALSE,
  nthreads = 1L,
  parallel = TRUE,
  refine.af = FALSE
)
}
\arguments{
\item{raw.data}{Numeric matrix (cells x detectors)}

\item{unmixed}{Numeric matrix (cells x fluors)}

\item{unmix.fun}{Unmixing function (fast OLS or WLS)}

\item{combined.spectra}{Numeric matrix (fluors x detectors)}

\item{weights}{Numeric vector (n detectors)}

\item{pos.thresholds}{Numeric vector (n fluors)}

\item{af.idx}{Integer vector (cells), AF spectrum assignments}

\item{af.spectra}{Numeric matrix (n af_variants x n detectors)}

\item{optimize.fluors}{Character vector of fluorophores present in variants}

\item{variants}{List of variant matrices per fluorophore}

\item{delta.list}{List of delta matrices per fluorophore}

\item{delta.norms}{List of delta norms per fluorophore}

\item{fluorophores}{Character vector of fluorophore names}

\item{af.idx.in.spectra}{Integer, row index of AF in spectra}

\item{asp}{The AutoSpectral parameter list.}

\item{k}{Integer, number of variants to test}

\item{cell.weighting}{Logical, use cell-specific weights}

\item{cell.weight.regularize}{Logical, regularize cell weights}

\item{nthreads}{Integer, number of threads}

\item{parallel}{Logical, whether to use parallel processing}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Parallel backend for per-cell spectral optimization in R.
}
