% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.r
\name{plot.convergence}
\alias{plot.convergence}
\title{Plot Convergence}
\usage{
\method{plot}{convergence}(convergence.log, asp)
}
\arguments{
\item{convergence.log}{Dataframe with convergence data of AutoSpectral.}

\item{asp}{List with AutoSpectral parameters.}
}
\value{
None. The function saves the generated convergence plot to a file.
}
\description{
Plots convergence of iterative refinement of the spillover matrix.
}
