% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_log_together.R
\name{plot.density.log.together}
\alias{plot.density.log.together}
\title{Plot Density of Log Values Together}
\usage{
\method{plot}{density.log.together}(x.table, x.label, plot.file.path, asp)
}
\arguments{
\item{x.table}{List of data frames containing values to be plotted.}

\item{x.label}{Label for the x-axis.}

\item{plot.file.path}{Path to save the plot file.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the density of log values for initial and
    final steps of AutoSpectral, and calculation with positive and negative
    populations, using ggplot2.
}
