% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gate_define.r
\name{plot.gate.define}
\alias{plot.gate.define}
\title{Plot Gate Define}
\usage{
\method{plot}{gate.define}(
  samp,
  gate.data,
  gate.marker,
  gate.bound,
  gate.region,
  gate.population,
  scatter.and.channel.label,
  asp
)
}
\arguments{
\item{samp}{Sample identifier.}

\item{gate.data}{Matrix containing gate data points.}

\item{gate.marker}{Vector containing gate marker names.}

\item{gate.bound}{List containing gate boundary information.}

\item{gate.region}{List containing gate region information.}

\item{gate.population}{List containing gate population information.}

\item{scatter.and.channel.label}{Named vector mapping scatter and channel labels.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the gate during the definition step,
    including boundaries and regions.
}
