% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Plot Heatmap}
\usage{
\method{plot}{heatmap}(matrix, asp, plot.prefix = NULL, number.labels)
}
\arguments{
\item{matrix}{Matrix or dataframe containing spectral data.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{plot.prefix}{Optional prefix for the plot filename.}

\item{number.labels}{Logical indicating whether to add number labels to the heatmap.}
}
\value{
Saves the heatmap plot as a JPEG file and the SSM data as a CSV file
    in the specified directory.
}
\description{
This function plots a matrix as a heatmap and saves it as a
    JPEG file.
}
