% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity_matrix.R
\name{plot.similarity.matrix}
\alias{plot.similarity.matrix}
\title{Plot Similarity Matrix}
\usage{
\method{plot}{similarity.matrix}(spectra, asp, plot.prefix = NULL)
}
\arguments{
\item{spectra}{Data frame containing spectral data.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{plot.prefix}{Optional prefix for the plot filename.}
}
\value{
Saves the heatmap plot as a JPEG file in the specified directory.
}
\description{
This function plots the similarity matrix as a heatmap and saves
    it as a JPEG file. It also calculates and displays the complexity index.
}
