% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spread_matrix.R
\name{plot.spread.matrix}
\alias{plot.spread.matrix}
\title{Plot Spillover Spread Matrix}
\usage{
\method{plot}{spread.matrix}(spectra, asp, plot.prefix = NULL)
}
\arguments{
\item{spectra}{Matrix containing spectral data. Fluorophores x detectors.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{plot.prefix}{Optional prefix for the plot filename.}
}
\value{
Saves the heatmap plot as a JPEG file and the SSM data as a CSV file
    in the specified directory.
}
\description{
This function plots the spillover spread matrix (SSM) as a heatmap
    and saves it as a JPEG file. It also saves the SSM data as a CSV file.
}
