% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bd_spectra.R
\name{read.bd.spectra}
\alias{read.bd.spectra}
\title{Extract Spectra From BD FCS File}
\usage{
read.bd.spectra(fcs.file)
}
\arguments{
\item{fcs.file}{Path and filename to the FCS file to be read.}
}
\value{
The spillover matrix as fluorophores x detectors, with fluorophore
and detector names applied.
}
\description{
Extracts the spectra (spillover) values from BD FCS files (e.g., A8, S8)
}
