% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_channel.r
\name{read.channel}
\alias{read.channel}
\title{Read Channel Information}
\usage{
read.channel(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{Directory containing control files.}

\item{control.def.file}{File containing control definitions.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A data frame containing the original and corrected channel names.
}
\description{
This function reads channel information from control files and
corrects channel names based on specified forbidden characters.
}
