% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scatter_parameter.r
\name{read.scatter.parameter}
\alias{read.scatter.parameter}
\title{Read Scatter Parameters}
\usage{
read.scatter.parameter(asp)
}
\arguments{
\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A vector of scatter parameters.
}
\description{
This function reads scatter parameters from a specified file or
    uses default scatter parameters if the file is not available.
}
