% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCSheader.R
\name{readFCSheader}
\alias{readFCSheader}
\title{Read FCS Header}
\usage{
readFCSheader(fcs.path, keyword = NULL)
}
\arguments{
\item{fcs.path}{A character string specifying the file path (directory and
file name) for the .fcs file to be read.}

\item{keyword}{Optional argument specifying which keyword(s) to return if the
whole set is not desired.}
}
\value{
A named list of keywords (metadata)
}
\description{
A light weight FCS file header reader based on \code{flowstate}, trying to mimic
\code{flowCore}.
}
\references{
Laniewski, Nathan. \emph{flowstate}.
\url{https://github.com/nlaniewski/flowstate}

Ellis B, Haaland P, Hahne F, Le Meur N, Gopalakrishnan N, Spidlen J, Jiang M,
Finak G (2025). \emph{flowCore: Basic structures for flow cytometry data}.
\doi{10.18129/B9.bioc.flowCore}
\url{https://bioconductor.org/packages/flowCore}
}
\seealso{
\code{\link[flowCore:read.FCSheader]{read.FCSheader}}
}
