% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_unmixing_lot_plot.r
\name{refine.unmixing.log.plot}
\alias{refine.unmixing.log.plot}
\title{Plot Density of Log Values}
\usage{
refine.unmixing.log.plot(x.data, x.label, plot.file.path, asp)
}
\arguments{
\item{x.data}{Data frame containing values to be plotted.}

\item{x.label}{Label for the x-axis.}

\item{plot.file.path}{Path to save the plot file.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the density of log values, excluding matrix
diagonal and zero values, using ggplot2.
}
