% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_unmixing.r
\name{refine.unmixing}
\alias{refine.unmixing}
\title{Refine Unmixing}
\usage{
refine.unmixing(
  spectra.initial,
  flow.control,
  asp,
  clean.expr = FALSE,
  plot.prefix = "Refined"
)
}
\arguments{
\item{spectra.initial}{Initial spectra matrix.}

\item{flow.control}{List containing flow control information.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{clean.expr}{Logical indicating whether to clean expression data.}

\item{plot.prefix}{Optional prefix for the plot title.}
}
\value{
The refined spectra matrix.
}
\description{
This function refines the unmixing process for spectral data,
    iteratively updating the spectra matrix and calculating unmixing errors.
}
