% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_flow_control.R
\name{reload.flow.control}
\alias{reload.flow.control}
\title{Reload Flow Control Information}
\usage{
reload.flow.control(control.def.file, control.dir, asp)
}
\arguments{
\item{control.def.file}{File containing control definitions.}

\item{control.dir}{Directory containing control files.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A list containing the reloaded flow control information.
}
\description{
This function reloads essential information from control files
    to permit rapid unmixing at a later date.
}
