% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_af_artefact_id.R
\name{run.af.artefact.id}
\alias{run.af.artefact.id}
\title{Run Autofluorescence Artefact Identification}
\usage{
run.af.artefact.id(clean.expr, negative.sample, spectral.channel, asp)
}
\arguments{
\item{clean.expr}{List containing cleaned expression data.}

\item{negative.sample}{Vector of negative sample names for which
autofluorescence artefact identification is to be performed.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A list containing the identified autofluorescence artefacts for each negative sample.
}
\description{
This function runs the identification of autofluorescence artefacts
    on a list of negative samples, using the specified parameters and settings.
}
