% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_peacoqc.R
\name{run.peacoQC}
\alias{run.peacoQC}
\title{Run PeacoQC}
\usage{
run.peacoQC(
  expr.data,
  spectral.channel,
  all.channels,
  asp,
  figures = TRUE,
  parallel = FALSE,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{expr.data}{A list containing the expression data for each sample.}

\item{spectral.channel}{A character vector specifying the spectral channels.}

\item{all.channels}{A character vector specifying all channels.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{figures}{Logical, if \code{TRUE} creates the main figures to show the
impact of intrusive autofluorescent event removal and scatter-matching for
the negatives.}

\item{parallel}{Logical, default is \code{FALSE}, in which case parallel processing
will not be used. Set to \code{TRUE} to run in parallel.}

\item{threads}{Number of cores to use for parallel processing, default is \code{1}.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A list containing the cleaned expression data for each sample.
}
\description{
This function runs PeacoQC to remove flow fluctuation errors from expression
data using parallel processing if specified.
}
