% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_universal_negative.R
\name{run.universal.negative}
\alias{run.universal.negative}
\title{Run Universal Negative}
\usage{
run.universal.negative(
  clean.expr,
  univ.sample,
  universal.negatives,
  scatter.param,
  peak.channels,
  downsample,
  negative.n,
  positive.n,
  spectral.channel,
  asp,
  control.type,
  scatter.match = TRUE,
  k.neighbors = 3L,
  intermediate.figures = FALSE,
  main.figures = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{clean.expr}{A matrix containing cleaned expression data.}

\item{univ.sample}{A character vector specifying the names of universal
negative samples.}

\item{universal.negatives}{A list containing universal negative control
parameters.}

\item{scatter.param}{A character vector specifying the scatter parameters.}

\item{peak.channels}{A character vector specifying the peak channels.}

\item{downsample}{A numeric value indicating the downsampling factor.}

\item{negative.n}{A numeric value indicating the number of negative events.}

\item{positive.n}{A numeric value indicating the number of positive events.}

\item{spectral.channel}{A character vector specifying the spectral channels.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{control.type}{A character string specifying the
type of control: \code{beads} or \code{cells}.}

\item{scatter.match}{Logical, default is \code{TRUE}. Whether to select negative
events based on scatter profiles matching the positive events.}

\item{k.neighbors}{Numeric, number of scatter-matched unstained events to
pair with every positive event for background determination. Default is \code{3}.}

\item{intermediate.figures}{Logical, if \code{TRUE} returns additional figures to
show the inner workings of the cleaning, including definition of low-AF cell
gates on the PCA-unmixed unstained and spectral ribbon plots of the AF
exclusion from the unstained. Default is \code{FALSE} to speed up processing.}

\item{main.figures}{Logical, if \code{TRUE} creates the main figures to show the
impact of intrusive autofluorescent event removal and scatter-matching for
the negatives.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A list containing the processed expression data for each universal
negative sample.
}
\description{
This function processes universal negative samples to generate expression
data based on specified parameters.
}
