% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_unmixing_matrix.R
\name{save.unmixing.matrix}
\alias{save.unmixing.matrix}
\title{Save Unmixing Matrix}
\usage{
save.unmixing.matrix(
  spectra,
  weights = NULL,
  filename = "unmixing_matrix",
  output.dir = "./table_spectra",
  figures = TRUE,
  color.palette = "mako"
)
}
\arguments{
\item{spectra}{Matrix or dataframe containing spectral data. Note that the
unmixing matrix depends on the combination of fluorophore signals present in
the spectral mixing matrix.}

\item{weights}{Optional numeric vector of weights, one per fluorescent
detector. Default is \code{NULL}, in which case OLS unmixing will be calculated.}

\item{filename}{Character string defining the output file name. Default is
\code{unmixing_matrix}.}

\item{output.dir}{Optional output directory. Default is \code{./table_spectra}.}

\item{figures}{Logical, if \code{TRUE}, produces a spectral heatmap of the unmixing
matrix in \code{output.dir}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{mako}. Options
are the viridis color options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis},
\code{cividis}, \code{rocket}, \code{mako} and \code{turbo}.}
}
\description{
This function calculates and saves the unmixing matrix to a CSV file. By
default, it will produce an ordinary least-squares (OLS) matrix. To save a
weighted least-squares (WLS) unmixing matrix, supply detector weights to
argument \code{weights}. Weights should be a numeric vector the same length as the
detectors in \code{spectra}. Good options for this are to use 1) detector noise
(reliability) measurements, for instance, CVs from instrument QC or 2) empirical
data from an FCS file. For the latter, take the raw data in the spectral
detector channels and apply function \code{colMeans}. This provides the average
signal level in the detecors, which we can assume is proportional to the noise
level. In both cases, the weights should be the inverse (1/noise). Use function
\code{calculate.weights()} to extract \code{colMeans} weighting from an FCS file.
}
