% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_match_plot.R
\name{scatter.match.plot}
\alias{scatter.match.plot}
\title{Plot Scatter-Matching of Universal Negative}
\usage{
scatter.match.plot(
  pos.expr.data,
  neg.expr.data,
  fluor.name,
  scatter.param,
  asp,
  color.palette = "rainbow"
)
}
\arguments{
\item{pos.expr.data}{A matrix containing the positive expression data.}

\item{neg.expr.data}{A matrix containing the negative expression data.}

\item{fluor.name}{A character string specifying the fluorophore name.}

\item{scatter.param}{A character vector specifying the scatter parameters.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{rainbow}, which will
be similar to FlowJo or SpectroFlo. Other pptions are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}
}
\value{
None. The function saves the generated scatter plot to a file.
}
\description{
This function generates scatter plots for matching positive and negative
expression data, selected based on scatter parameters gates.
}
