% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_matrix_plot.R
\name{similarity.matrix.plot}
\alias{similarity.matrix.plot}
\title{Similarity Matrix Plot}
\usage{
similarity.matrix.plot(spectra, asp, plot.prefix = NULL)
}
\arguments{
\item{spectra}{Data frame containing spectral data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{plot.prefix}{Optional prefix for the plot filename. Default is \code{NULL}}
}
\value{
Saves the heatmap plot as a JPEG file in the specified directory.
}
\description{
This function plots the similarity matrix as a heatmap and saves
it as a JPEG file. It also calculates and displays the complexity index.
}
