% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_heatmap.R
\name{spectral.heatmap}
\alias{spectral.heatmap}
\title{Spectral Heatmap}
\usage{
spectral.heatmap(
  spectra,
  asp,
  plot.prefix = NULL,
  legend.label = "Intensity",
  output.dir = NULL
)
}
\arguments{
\item{spectra}{Matrix or dataframe containing spectral data
format: fluorophores x detectors.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{plot.prefix}{Optional prefix for the plot filename.}

\item{legend.label}{Character string that will appear on the heatmap legend.}

\item{output.dir}{Optional output directory. Default is NULL, in which case
the spectra figure folder will be used.}
}
\value{
Saves the heatmap plot as a JPEG file in the specified directory.
}
\description{
This function plots a spectral matrix as a heatmap and saves it
as a JPEG file.
}
