% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_ribbon_plot.R
\name{spectral.ribbon.plot}
\alias{spectral.ribbon.plot}
\title{Spectral Ribbon Plot}
\usage{
spectral.ribbon.plot(
  pos.expr.data,
  neg.expr.data,
  spectral.channel,
  asp,
  fluor.name,
  plot.prefix = NULL,
  af = FALSE,
  removed.data = NULL,
  figure.dir = NULL,
  factor.names = NULL
)
}
\arguments{
\item{pos.expr.data}{A matrix containing the positive expression data.}

\item{neg.expr.data}{A matrix containing the negative expression data.}

\item{spectral.channel}{A character vector specifying the spectral channels.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{fluor.name}{A character string specifying the fluorophore name.}

\item{plot.prefix}{A character string to prefix the plot file name.
Default is \code{NULL}}

\item{af}{A logical value indicating whether autofluorescence removal is
being performed. Default is \code{FALSE}}

\item{removed.data}{A matrix containing the removed data, if applicable.
Default is \code{NULL}. If omitted, only two groups (facets) are plotted.}

\item{figure.dir}{Output folder where the figures will be created. Default is
\code{NULL}, enabling automatic selection inside AutoSpectral.}

\item{factor.names}{Optional titles for the facets on the plot. Default is
\code{NULL}, enabling automatic selection inside AutoSpectral. A character vector
containing three elements should be provided unless \code{af} is set to \code{TRUE} and
\code{removed.data} is \code{NULL}, in which case two labels must be provided.}
}
\value{
None. The function saves the generated spectral ribbon plot to
a file.
}
\description{
This function generates spectral ribbon plots for positive and negative
expression data.
}
