% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_variant_plot.R
\name{spectral.variant.plot}
\alias{spectral.variant.plot}
\title{Spectral Variant Plot}
\usage{
spectral.variant.plot(
  spectra.variants,
  median.spectrum,
  title = "Spectral_variants",
  save = FALSE,
  plot.width = NULL,
  plot.height = 5,
  plot.dir = "./figure_spectral_variants",
  variant.fill.color = "red",
  variant.fill.alpha = 0.7,
  median.line.color = "black",
  median.linewidth = 1
)
}
\arguments{
\item{spectra.variants}{A matrix of variations in the normalized spectra
signature for a fluorophore, rows being the variants, columns being the
detectors.}

\item{median.spectrum}{The optimized reference spectrum for the fluorophore,
usually obtained from spectra prepared using AutoSpectral via
\code{get.fluorophore.spectra} on cleaned control data.}

\item{title}{Optional title to pass to the plot and output filename. Default is
\code{Spectral_variants}.}

\item{save}{Logical, controls whether the plot is displayed to the Viewer
(\code{FALSE}) or also saved to disk (\code{TRUE}).}

\item{plot.width}{Optional numeric to control saved plot width. Default is
\code{NULL}, in which case the plot width will be calculated based on the number
of detectors in the data.}

\item{plot.height}{Optional numeric to control saved plot width. Default is \code{5}.}

\item{plot.dir}{Directory where the files will be saved if \code{save = TRUE}.
Default is \code{./figure_spectral_variants}.}

\item{variant.fill.color}{Color for the shaded region indicating the range of
variation in the spectra. Feeds to \code{fill} in \code{geom_ribbon}. Default is "red".}

\item{variant.fill.alpha}{Transparency (alpha) for the color in
\code{variant.fill.color}. How intense the color of the variant spectra will be.
Default is \code{0.7}}

\item{median.line.color}{Color for the line representing the median or
optimized single spectrum. Default is "black".}

\item{median.linewidth}{Width of the line for the single optimized spectrum.
Default is \code{1}.}
}
\value{
Plots are displayed to the Viewer. Files are saved if saving enabled.
}
\description{
This function plots a spectral ribbon showing the optimized fluorophore
signature with the range of variation from the spectral variants.
}
