% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_extreme_events.R
\name{trim.extreme.events}
\alias{trim.extreme.events}
\title{Trim Extreme Events}
\usage{
trim.extreme.events(expr.data, peak.channel, trim.factor)
}
\arguments{
\item{expr.data}{A  matrix containing the expression data.}

\item{peak.channel}{A character string specifying the peak channel to be
used for trimming.}

\item{trim.factor}{A numeric value indicating the proportion of extreme
events to trim from both ends of the peak channel distribution.}
}
\value{
A matrix with the extreme events trimmed.
}
\description{
This function trims extreme events from expression data based
on a specified peak channel and trim factor.
}
