% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_gate.R
\name{tune.gate}
\alias{tune.gate}
\title{Tune Landmark Gate}
\usage{
tune.gate(
  control.file,
  control.dir,
  asp,
  gate.name,
  n.cells = c(100, 500, 2000),
  percentiles = c(30, 50, 70),
  grid.n = 100,
  bandwidth.factor = 1,
  fsc.channel = NULL,
  ssc.channel = NULL,
  fsc.lims = NULL,
  ssc.lims = NULL,
  output.dir = "./figure_gate_tuning",
  filename = "gate_tuning",
  color.palette = "mako",
  boundary.color = "red",
  points.to.plot = 1e+05,
  width = 9,
  height = 10
)
}
\arguments{
\item{control.file}{File path and name for the CSV file defining the single-
color control file names, fluorophores they represent, marker names, peak
channels, and gating requirements.}

\item{control.dir}{File path to the single-stained control FCS files.}

\item{asp}{The AutoSpectral parameter list defined using
\code{get.autospectral.param}.}

\item{gate.name}{Character, name for the gate. Useful for distinguishing gates
when you have multiple types. Must match one string (name) in the \code{gate.name}
column of the \code{control.file}.}

\item{n.cells}{The number of cells to use for defining the gate boundary. The
default is \code{c(100, 500, 2000)} and will test each of those values. This many
cells will be selected from the peak channel (brightest first) in the single-
color controls. For example, if you set \code{200} and marked files such as
\code{CD3-PE.fcs} and \code{CD19-FITC.fcs} as \code{gate.define=TRUE} in the control file,
the brightest 200 events in the YG1 channel from the CD3-PE file and the
brightest 200 events in the B1 channel for the CD19-FITC file would be used
to define the gate.}

\item{percentiles}{Numeric 1 - 100, default \code{c(30, 50, 70)}. The percentile
cutoffs to test for density in the scatter to use for defining the gate. For
example, a value of \code{50} would take the 50\% of cells closest to the density
peak, more or less. Smaller numbers will define a tighter gate.}

\item{grid.n}{Numeric, default \code{100}. The binning grid for the kernel density
estimation. If \code{n.cells} is very low, you may wish to lower this number to
compress the search space.}

\item{bandwidth.factor}{Numeric, default \code{1}. A multiplier for the bandwidth
for the kernel density estimation. Larger numbers will smooth the density,
reducing discrimination between peaks in the density (such as between live
and dead cells).}

\item{fsc.channel}{Channel to use for Forward Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[1]}, which is appropriate for your
machine.}

\item{ssc.channel}{Channel to use for Side Scatter. Default \code{NULL} will
use the \code{asp$default.scatter.parameter[2]}, which is appropriate for your
machine. For machines with multiple side scatter measurements, you can change
this.}

\item{fsc.lims}{Numeric vector. Limits for plotting the FSC. The default \code{NULL}
uses c\code{c(asp$scatter.data.min.x, asp$scatter.data.max.x)}.}

\item{ssc.lims}{Numeric vector. Limits for plotting the SSC. The default \code{NULL}
uses c\code{c(asp$scatter.data.min.y, asp$scatter.data.max.y)}.}

\item{output.dir}{File path where you want to save the results. Default is
\code{./figure_gate_tuning}.}

\item{filename}{Character, name for the output files. Default is \code{gate_tuning}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{plasma}. Use \code{rainbow}
to be similar to FlowJo or SpectroFlo. Other options are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{boundary.color}{Color for the gate boundary line on the plot. Default
is \code{darkgoldenrod1}.}

\item{points.to.plot}{Numeric, default \code{1e5}. Maximum number of points to show
on the plot. More points will take longer, but you really shouldn't have even
close to this number when defining the gate with landmarks.}

\item{width}{Numeric, default \code{4}. Width of the saved plot.}

\item{height}{Numeric, default \code{4}. Height of the saved plot.}
}
\value{
A combined plot of all gates. Saves plots of the gates defined using
combinations of the specified parameters.
}
\description{
This function provides a parameter search grid for optimizing FSC/SSC gating
on the single-stained control files. Set up the control file and specify gate
names (for example, group controls with similar expected scatter regions into
one "gate.name"). \code{tune.gate()} selects the brightest \code{n.cells} positive events
from each file linked to a given \code{gate.name}, and cuts the density peak at
\code{percentiles}. You get plots showing the output gate boundaries plotted in the
\code{boundary.color}. This uses the "landmark" gating system. Note that for this
to work, you must supply some controls with staining. Unstained samples will
be excluded, and if you only supply unstained samples, it will stop. To draw
gates on unstained samples, use the "density" method: see \code{define.gate.density()}.
}
\references{
Laniewski, Nathan. \emph{flowstate}.
\url{https://github.com/nlaniewski/flowstate}
}
\seealso{
\itemize{
\item \code{\link[=define.gate.landmarks]{define.gate.landmarks()}}
\item \code{\link[=define.gate.density]{define.gate.density()}}
\item \code{\link[=gate.sample.plot]{gate.sample.plot()}}
}
}
