% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unmix_fix.R
\name{unmix.fix.plot}
\alias{unmix.fix.plot}
\title{Unmixing Fixing Plot}
\usage{
unmix.fix.plot(
  original.data,
  compensated.data,
  plot.idx,
  asp,
  unstained.thresholds,
  unstained.margin,
  spread.estimate,
  output.dir
)
}
\arguments{
\item{original.data}{The fully stained sample with the original unmixing.}

\item{compensated.data}{The fully stained sample with compensation corrections
applied.}

\item{plot.idx}{The index of the biplot comparisons to be plotted.}

\item{asp}{The AutoSpectral parameter list.}

\item{unstained.thresholds}{Numeric between 0 and 1, default 0.99. The threshold
used to determine positivity in the fully stained sample (the percentile
on the unstained sample in that channel).}

\item{unstained.margin}{Numeric, default 1.5. The fudge factor above the
unstained.threshold.}

\item{spread.estimate}{Numeric matrix. The estimated spillover spread for the
panel.}

\item{output.dir}{The folder where the plot will be saved.}
}
\value{
No returns. Creates a jpeg plot.
}
\description{
Plots the data before and after automated unmixing corrections.
}
