% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_ols_fast.R
\name{unmix.ols.fast}
\alias{unmix.ols.fast}
\title{Unmix OLS Fast}
\usage{
unmix.ols.fast(raw.data, spectra, weights = NULL)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{weights}{Dummy argument to allow dynamic switching between OLS and WLS.
Default is \code{NULL}. Values passed to \code{weights} will be ignored.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Faster solver for per-cell optimization workflow. Performs spectral unmixing
using ordinary least squares.
}
