% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_ols.R
\name{unmix.ols}
\alias{unmix.ols}
\title{unmix.ols}
\usage{
unmix.ols(raw.data, spectra)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Performs spectral unmixing using ordinary least squares
}
