% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_wls_fast.R
\name{unmix.wls.fast}
\alias{unmix.wls.fast}
\title{Unmix WLS Fast}
\usage{
unmix.wls.fast(raw.data, spectra, weights = NULL)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{weights}{The weighting values for the detectors. No checks are performed
in this function; \code{unmix.wls} should be used for most cases.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Faster solver for per-cell optimization workflow. Performs spectral unmixing
using weighted least squares.
}
