% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_flowjo_mtx.R
\name{write.flowjo.mtx}
\alias{write.flowjo.mtx}
\title{Write FlowJo Matrix}
\usage{
write.flowjo.mtx(spillover.matrix, flowjo.file.name)
}
\arguments{
\item{spillover.matrix}{The spillover matrix (fluorophores x fluorophores).}

\item{flowjo.file.name}{The path and filename given to the generated FlowJo
compensation file.}
}
\value{
No return. Write the xml .mtx file.
}
\description{
A quick and dirty method for writing a FlowJo-compatible compensation matrix
file from the spillover matrix.
}
