% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFCS.R
\name{writeFCS}
\alias{writeFCS}
\title{Write FCS File}
\usage{
writeFCS(mat, keys, file.name, output.dir)
}
\arguments{
\item{mat}{The expression data, containing both unmixed data and any
retained parameters such as scatter and time.}

\item{keys}{The keywords to be written to the file.}

\item{file.name}{The name of the FCS file to be written.}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file.}
}
\description{
A light weight FCS file binary writer based on \code{flowstate}. Lower memory usage
and faster than \code{flowCore}.
}
\references{
Laniewski, Nathan. \emph{flowstate}.
\url{https://github.com/nlaniewski/flowstate}

Ellis B, Haaland P, Hahne F, Le Meur N, Gopalakrishnan N, Spidlen J, Jiang M,
Finak G (2025). \emph{flowCore: Basic structures for flow cytometry data}.
\doi{10.18129/B9.bioc.flowCore}
\url{https://bioconductor.org/packages/flowCore}
}
\seealso{
\code{\link[flowCore:write.FCS]{write.FCS}}
}
