% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_control_file.R
\name{check.control.file}
\alias{check.control.file}
\title{Check Control File}
\usage{
check.control.file(control.dir, control.def.file, asp, strict = FALSE)
}
\arguments{
\item{control.dir}{File path to the single-stained control FCS files.}

\item{control.def.file}{CSV file defining the single-color control file names,
fluorophores they represent, marker names, peak channels, and gating requirements.}

\item{asp}{The AutoSpectral parameter list defined
using \code{get.autospectral.param}.}

\item{strict}{Logical. Controls whether the function triggers a break or
continues and outputs a list of errors. Default is \code{FALSE}.}
}
\value{
A dataframe of errors and warnings intended to help the user fix
problems with the \code{control.def.file}.
}
\description{
Attempts to find potential failure points and input errors in the control file
\code{control.def.file} used to define the single-stained control setup for
AutoSpectral.
}
