% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gates.R
\name{check.gates}
\alias{check.gates}
\title{Check Gates For Errors}
\usage{
check.gates(gate.list, control.table, asp)
}
\arguments{
\item{gate.list}{Named list of gates. To use this, pre-define the
gates using \code{define.gate.landmarks()} and/or \code{define.gate.density()}, ensure
that the names of the gates correspond to the names in the \code{control.def.file},
and ensure that the \code{gate.name} column has been filled in for the
\code{control.def.file}.}

\item{control.table}{Dataframe or table of the control file, read in via
\code{define.flow.control()} and cleaned up by that function.}

\item{asp}{The AutoSpectral parameter list defined using \code{get.autospectral.param}.}
}
\value{
Silently returns \code{TRUE} if all checks pass. If any check fails, the
pipeline halts.
}
\description{
A small helper function to check for non-standard or crash-inducing input when
users supply gates to \code{define.flow.control()}.
}
\seealso{
\itemize{
\item \code{\link[=tune.gate]{tune.gate()}}
\item \code{\link[=define.gate.landmarks]{define.gate.landmarks()}}
\item \code{\link[=define.gate.density]{define.gate.density()}}
\item \code{\link[=do.gate]{do.gate()}}
\item \code{\link[=define.flow.control]{define.flow.control()}}
}
}
