% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_controls.R
\name{clean.controls}
\alias{clean.controls}
\title{Clean Controls}
\usage{
clean.controls(
  flow.control,
  asp,
  time.clean = FALSE,
  trim = FALSE,
  trim.factor = NULL,
  af.remove = FALSE,
  universal.negative = TRUE,
  downsample = TRUE,
  negative.n = asp$negative.n,
  positive.n = asp$positive.n,
  scatter.match = TRUE,
  refactor = TRUE,
  return.separately = FALSE
)
}
\arguments{
\item{flow.control}{A list prepared using \code{define.flow.control}, containing
the data and essential information about the cytometer and data structure.}

\item{asp}{The AutoSpectral parameter list, prepared using
\code{get.autospectral.param}.}

\item{time.clean}{Logical, default is \code{FALSE}. Whether to run PeacoQC to
remove time-based inconsistencies in the controls.}

\item{trim}{Logical, default is \code{FALSE}. Whether to remove extreme events
(positive and negative) from controls.}

\item{trim.factor}{Numeric. Default is \code{asp$rlm.trim.factor}. Required if
\code{trim = TRUE}.}

\item{af.remove}{Logical, default is \code{FALSE}. Whether to remove intrusive
autofluorescence contamination from cell controls using PCA-based
identification and gating. Requires universal negatives to be defined in the
control file and in \code{flow.control}.}

\item{universal.negative}{Logical, default is \code{TRUE}. Whether to use a
universal negative sample as the negative for spectral extraction. Requires
universal negatives to be defined in the control file and in \code{flow.control}.}

\item{downsample}{Logical, default is \code{TRUE}. Whether to reduce cell and bead
control events to speed up processing.}

\item{negative.n}{Integer. Number of events to include in the downsampled
negative population. Default is \code{asp$negative.n}.}

\item{positive.n}{Integer. Number of events to include in the downsampled
positive population. Default is \code{asp$positive.n}.}

\item{scatter.match}{Logical, default is \code{TRUE}. Whether to select negative
events based on scatter profiles matching the positive events. Defines a
region of FSC and SSC based on the distribution of selected positive events.}

\item{refactor}{Logical, default is \code{TRUE}. Reassigns event indices after
cleaning steps to ensure consistent indexing.}

\item{return.separately}{Logical, default is \code{FALSE}. If \code{TRUE}, returns the
cleaned expression data as output in addition to storing it in
\code{flow.control}.}
}
\value{
If \code{return.separately = TRUE}, returns the cleaned expression data.
Otherwise, no return value; the cleaned data is stored in \code{flow.control} for
downstream use.
}
\description{
A four-part function to clean single-color controls in order to extract
fluorophore signatures. Any part can be run independently:
\itemize{
\item \strong{Stage 1}: PeacoQC to eliminate flow artefacts.
\item \strong{Stage 2}: Trimming to eliminate extreme events.
\item \strong{Stage 3}: Autofluorescence noise removal using PCA unmixing on matching
unstained (cells only).
\item \strong{Stage 4}: Brightest event selection from positive, universal negative
from matching negative, and downsampling to speed up RLM spectra
optimization.
}
}
