% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_controls.R
\name{clean.controls}
\alias{clean.controls}
\title{Clean Controls}
\usage{
clean.controls(
  flow.control,
  asp,
  time.clean = FALSE,
  trim = FALSE,
  trim.factor = NULL,
  af.remove = TRUE,
  universal.negative = TRUE,
  downsample = TRUE,
  negative.n = asp$negative.n,
  positive.n = asp$positive.n,
  scatter.match = TRUE,
  scrub = FALSE,
  intermediate.figures = FALSE,
  main.figures = TRUE,
  parallel = FALSE,
  verbose = TRUE,
  threads = NULL
)
}
\arguments{
\item{flow.control}{A list prepared using \code{define.flow.control}, containing
the data and essential information about the cytometer and data structure.}

\item{asp}{The AutoSpectral parameter list, prepared using
\code{get.autospectral.param}.}

\item{time.clean}{Logical, default is \code{FALSE}. Whether to run PeacoQC to
remove time-based inconsistencies in the controls.}

\item{trim}{Logical, default is \code{FALSE}. Whether to remove extreme events
(positive and negative) from controls.}

\item{trim.factor}{Numeric. Default is \code{asp$rlm.trim.factor}. Required if
\code{trim = TRUE}.}

\item{af.remove}{Logical, default is \code{TRUE}. Whether to remove intrusive
autofluorescence contamination from cell controls using PCA-based
identification and gating. Requires universal negatives to be defined in the
control file and in \code{flow.control}.}

\item{universal.negative}{Logical, default is \code{TRUE}. Whether to use a
universal negative sample as the negative for spectral extraction. Requires
universal negatives to be defined in the control file and in \code{flow.control}.}

\item{downsample}{Logical, default is \code{TRUE}. Whether to reduce cell and bead
control events to speed up processing.}

\item{negative.n}{Integer. Number of events to include in the downsampled
negative population. Default is \code{asp$negative.n}.}

\item{positive.n}{Integer. Number of events to include in the downsampled
positive population. Default is \code{asp$positive.n}.}

\item{scatter.match}{Logical, default is \code{TRUE}. Whether to select negative
events based on scatter profiles matching the positive events. Defines a
region of FSC and SSC based on the distribution of selected positive events.}

\item{scrub}{Logical, if \code{TRUE} allows for re-cleaning of already cleaned
data, provided there are clean data in \code{flow.control}.}

\item{intermediate.figures}{Logical, if \code{TRUE} returns additional figures to
show the inner workings of the cleaning, including definition of low-AF cell
gates on the PCA-unmixed unstained and spectral ribbon plots of the AF
exclusion from the unstained. Default is \code{FALSE} to speed up processing.}

\item{main.figures}{Logical, if \code{TRUE} creates the main figures to show the
impact of intrusive autofluorescent event removal and scatter-matching for
the negatives.}

\item{parallel}{Logical, default is \code{FALSE}, in which case parallel processing
will not be used. Parallel processing will likely be faster when many small
files are read in. If the data is larger, parallel processing may not
accelerate the process much.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}

\item{threads}{Numeric, number of threads to use for parallel processing.
Default is \code{NULL} which will revert to \code{asp$worker.process.n} if
\code{parallel=TRUE}.}
}
\value{
Returns a modified \code{flow.control} with the original data intact. New, cleaned
data and corresponding factors are stored in new slots.
}
\description{
A four-part function to clean single-color controls in order to extract
fluorophore signatures. Any part can be run independently:
\itemize{
\item \strong{Stage 1}: PeacoQC to eliminate flow artefacts. Not required in most cases.
\item \strong{Stage 2}: Trimming to eliminate extreme events. Not recommended for most
use cases.
\item \strong{Stage 3}: Autofluorescence noise removal using PCA unmixing on matching
unstained (cells only).
\item \strong{Stage 4}: Brightest event selection from positive, universal negative
from matching negative, and downsampling to speed up RLM spectra optimization.
}
}
