% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity_plot.R
\name{cosine.similarity.plot}
\alias{cosine.similarity.plot}
\title{Cosine Similarity Plot}
\usage{
cosine.similarity.plot(
  spectra,
  filename = "autospectral_similarity_matrix",
  plot.prefix = NULL,
  output.dir = "figure_similarity_heatmap",
  figure.width = 8,
  figure.height = 6,
  color.palette = "viridis"
)
}
\arguments{
\item{spectra}{Data frame or matrix containing spectral data.}

\item{filename}{Character string for the output file. Default is
\code{autospectral_similarity_matrix}.}

\item{plot.prefix}{Optional prefix for the plot filename. Default is \code{NULL}}

\item{output.dir}{File path where the plot will be created. Default is
\code{figure_similarity_heatmap}. The directory will be created if it does not
already exist.}

\item{figure.width}{Numeric. Width of output plot. Default is \code{8}.}

\item{figure.height}{Numeric. Height of output plot. Default is \code{6}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{viridis}. Options
are the viridis color options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis},
\code{cividis}, \code{rocket}, \code{mako} and \code{turbo}.}
}
\value{
Saves the heatmap plot as a JPEG file in the similarity directory.
}
\description{
This function plots the matrix of cosine similarities (AKA "Similarity Matrix")
as a heatmap and saves it as a JPEG file. It also calculates and displays the
mixing matrix condition number (AKA "Complexity Index") of the matrix.
}
