% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine.similarity}
\alias{cosine.similarity}
\title{Calculate cosine similarity}
\usage{
cosine.similarity(spectra.t)
}
\arguments{
\item{spectra.t}{The transposed spectral matrix (or dataframe), represented as
detectors in rows and fluorophores in columns.}
}
\value{
The cosine similarity matrix in fluorophores x fluorophores.
}
\description{
Calculates the cosine similarity between columns of the input matrix (spectra).
}
