% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biplot.R
\name{create.biplot}
\alias{create.biplot}
\title{Create Biplot}
\usage{
create.biplot(
  plot.data,
  x.dim,
  y.dim,
  asp,
  x.lab = NULL,
  y.lab = NULL,
  x.min = -5000,
  x.max = asp$expr.data.max,
  y.min = -5000,
  y.max = asp$expr.data.max,
  x.width.basis = -1000,
  y.width.basis = -1000,
  max.points = 1e+05,
  color.palette = "rainbow",
  save = TRUE,
  title = NULL,
  output.dir = NULL,
  width = 5,
  height = 5
)
}
\arguments{
\item{plot.data}{A matrix or dataframe containing the flow cytometry data to
be plotted. Column names should match the dimensions specified by \code{x.dim} and
\code{y.dim}.}

\item{x.dim}{String specifying the column of \code{plot.data} for the x-axis of
the plot.}

\item{y.dim}{String specifying the column of \code{plot.data} for the y-axis of
the plot.}

\item{asp}{The AutoSpectral parameter list.}

\item{x.lab}{An optional label for the x-axis. If none is given (default
\code{NULL}), the column name specified by \code{x.dim} will be used.}

\item{y.lab}{An optional label for the y-axis. If none is given (default
\code{NULL}), the column name specified by \code{y.dim} will be used.}

\item{x.min}{Minimum value for the x-axis. Default is \code{-5000}.}

\item{x.max}{Maximum value for the x-axis. Default is the value specified by
asp$expr.data.max, which will be the maximum for the cytometer.}

\item{y.min}{Minimum value for the y-axis. Default is \code{-5000}.}

\item{y.max}{Maximum value for the y-axis. Default is the value specified by
asp$expr.data.max, which will be the maximum for the cytometer.}

\item{x.width.basis}{Width basis for the biexponential transform for the
x-axis. Default is \code{-1000}.}

\item{y.width.basis}{Width basis for the biexponential transform for the
x-axis. Default is \code{-1000}.}

\item{max.points}{Number of points to plot (speeds up plotting). Default is
\code{1e5}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{rainbow}, which will
be similar to FlowJo or SpectroFlo. Other pptions are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{save}{Logical, if \code{TRUE}, saves a JPEG file to the \code{output.dir}.
Otherwise, the plot will simply be created in the Viewer.}

\item{title}{Optional title for the plot filename. If \code{NULL}, defaults to
\code{x.lab} vs. \code{y.lab}.}

\item{output.dir}{Optional output directory. Default is NULL, in which case
the current working directory will be used.}

\item{width}{Numeric, width of the saved plot. Default is \code{5}.}

\item{height}{Numeric, height of the saved plot. Default is \code{5}.}
}
\value{
Creates a biplot in the Viewer and optionally saves it as a JPEG file.
}
\description{
Create Biplot
}
