% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control_file.R
\name{create.control.file}
\alias{create.control.file}
\title{Create Control File}
\usage{
create.control.file(
  control.dir,
  asp,
  fill.gate.name = TRUE,
  filename = "fcs_control_file",
  legacy = FALSE,
  output.dir = getwd()
)
}
\arguments{
\item{control.dir}{file path to the single stained control fcs files}

\item{asp}{The AutoSpectral parameter list. Generate using
\code{get.autospectral.param}}

\item{fill.gate.name}{Logical, default is \code{TRUE}. Will attempt to automatically
assign gate names for the \code{gate.name} column if \code{TRUE}.}

\item{filename}{Character string defining the output filename. Default is
"fcs_control_file", to which .csv will be appended.}

\item{legacy}{Logical. If \code{FALSE}, gating-related columns will not be created
and the control file will be suitable only for the new automated spectral
extraction pipeline using \code{get.spectra.automated()}. To use the version 1
"legacy" pipeline for the extraction of fluorophore spectra, using gating and
\code{define.flow.control()}, set \code{legacy=TRUE}.}

\item{output.dir}{location where the CSV file will be written. Default is the
current working directory.}
}
\value{
No returns. Outputs a csv file called fcs_control_file.csv
}
\description{
A helper function to draft a description of your single stained control
files such that AutoSpectral can understand and process them correctly.
Given a set of single stained control fcs files, \code{create.control.file} will
produce a csv file listing the matching peak detector channels for your
fluorophores (if known). If your files contain bead or cell tags in the filename,
it will assign your controls as cells or beads, and will automatically fill in
the matching \code{universal.negative} filename when exactly one unstained/negative
control shares that control.type. \code{universal.negative} is left blank when
\code{control.type} could not be determined or when more than one unstained/negative
control of that type exists (in which case you will need to set it manually).
You will need to fill in any "No Match" results manually.
You will need to add marker names manually.
}
