% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control_file.R
\name{create.control.file}
\alias{create.control.file}
\title{Create Control File}
\usage{
create.control.file(control.dir, asp)
}
\arguments{
\item{control.dir}{file path to the single stained control fcs files}

\item{asp}{The AutoSpectral parameter list. Generate using
\code{get.autospectral.param}}
}
\value{
No returns. Outputs a csv file called fcs_control_file.csv
}
\description{
A helper function to draft a description of your single stained control
files such that AutoSpectral can understand and process them correctly.
Given a set of single stained control fcs files, \code{create.control.file} will
produce a csv file listing the matching peak detector channels for your
fluorophores (if known). If your files contain bead or cell tags in the filename,
it will assign your controls as cells or beads. You will need to fill in any
"No Match" results manually. You will need to set universal negatives manually.
You will need to add marker names manually.
}
