% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{create.heatmap}
\alias{create.heatmap}
\title{Create Heatmap Plot}
\usage{
create.heatmap(
  matrix,
  number.labels,
  plot.prefix = NULL,
  legend.label = "heatmap",
  triangular = FALSE,
  output.dir = NULL,
  fixed.scale = FALSE,
  scale.min = NULL,
  scale.max = NULL,
  color.palette = "viridis",
  figure.width = 8,
  figure.height = 6
)
}
\arguments{
\item{matrix}{Matrix or dataframe containing spectral data.}

\item{number.labels}{Logical indicating whether to add number labels to
the heatmap.}

\item{plot.prefix}{Optional prefix for the plot filename. Default is \code{NULL},
in which case the file will just be called \code{heatmap.jpg}}

\item{legend.label}{Character string that will appear on the heatmap legend.
Default is \code{heatmap}}

\item{triangular}{Logical. Plot the lower triangle of the matrix only,
diagonal included. Default is \code{FALSE}.}

\item{output.dir}{Optional output directory. Default is \code{NULL}, in which case
the working directory will be used.}

\item{fixed.scale}{Logical, determines whether to use an externally supplied
fixed scale (min and max) for the heatmap color scale. Useful for putting
multiple plots on the same scale. Default is \code{FALSE}}

\item{scale.min}{Optional numeric. Minimum for the fixed color scale.
Default is \code{NULL}, for no fixed scaling.}

\item{scale.max}{Optional numeric. Maximum for the fixed color scale.
Default is \code{NULL}, for no fixed scaling.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{viridis}. Options
are the viridis color options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis},
\code{cividis}, \code{rocket}, \code{mako} and \code{turbo}.}

\item{figure.width}{Numeric. Width of the heatmap figure. Default is \code{8}.}

\item{figure.height}{Numeric. Height of the heatmap figure. Default is \code{6}.}
}
\value{
Saves the heatmap plot as a JPEG file and the SSM data as a CSV file
in the specified directory.
}
\description{
This function plots a matrix as a heatmap and saves it as a JPEG file.
}
