% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_flow_control.r
\name{define.flow.control}
\alias{define.flow.control}
\title{Define Flow Control}
\usage{
define.flow.control(
  control.dir,
  control.def.file,
  asp,
  gate = TRUE,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{control.dir}{File path to the single-stained control FCS files.}

\item{control.def.file}{CSV file defining the single-color control file names,
fluorophores they represent, marker names, peak channels, and gating requirements.}

\item{asp}{The AutoSpectral parameter list defined using
\code{get.autospectral.param}.}

\item{gate}{Logical, default is \code{TRUE}, in which case, automated gating will
be performed. If \code{FALSE}, the FCS files will be imported without automatically
generated gates applied. That is, all data in the files will be used. This is
intended to allow the user to pre-gate the files in commercial software.}

\item{parallel}{Logical, default is \code{FALSE}, in which case parallel processing
will not be used. Parallel processing will likely be faster when many small
files are read in. If the data is larger, parallel processing may not
accelerate the process much.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A list (\code{flow.control}) with the following components:
\itemize{
\item \code{filename}: Names of the single-color control files.
\item \code{fluorophore}: Corresponding fluorophores used in the experiment.
\item \code{antigen}: Corresponding markers used in the experiment.
\item \code{control.type}: Type of control used (beads or cells).
\item \code{universal.negative}: Corresponding universal negative for each control.
\item \code{viability}: Logical factor; whether a control represents a viability marker.
\item \code{large.gate}: Logical factor; large gate setting.
\item \code{autof.channel.idx}: Index of the autofluorescence marker channel.
\item \code{event.number.width}: Width of the event number.
\item \code{expr.data.max}: Maximum expression data value.
\item \code{expr.data.max.ceil}: Ceiling of the maximum expression data value.
\item \code{expr.data.min}: Minimum expression data value.
\item \code{channel}: Preliminary peak channels for the fluorophores.
\item \code{channel.n}: Number of channels.
\item \code{spectral.channel}: Spectral channel information.
\item \code{spectral.channel.n}: Number of spectral channels.
\item \code{sample}: Sample names (fluorophores).
\item \code{scatter.and.channel}: FSC, SSC, and peak channel information.
\item \code{scatter.and.channel.label}: Labels for scatter and channel.
\item \code{scatter.and.channel.spectral}: FSC, SSC, and spectral channels.
\item \code{scatter.parameter}: Scatter parameters used for gating.
\item \code{event}: Event factor.
\item \code{event.n}: Number of events.
\item \code{event.sample}: Sample information for events. Links events to samples.
\item \code{event.type}: Type of events.
\item \code{expr.data}: Expression data used for extracting spectra.
}
}
\description{
A complex function designed to convert the single-stained control FCS files
and the metadata \code{control.def.file} into a data structure for AutoSpectral.
It reads the metadata file, locates the corresponding FCS files, determines
the gates needed based on variables such as beads or cells, \code{large.gate}, and
\code{viability.gate}, and then creates gates for each combination. It imports
and gates the data in the FCS files and assigns various factors to track
the data.
}
