% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate_af.r
\name{do.gate.af}
\alias{do.gate.af}
\title{Perform Gating on Autofluorescence Parameters}
\usage{
do.gate.af(gate.data, samp, asp, intermediate.figures = FALSE)
}
\arguments{
\item{gate.data}{A data frame containing the gate data.}

\item{samp}{A sample identifier.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{intermediate.figures}{Logical, if \code{TRUE} returns additional figures to
show the inner workings of the cleaning, including definition of low-AF cell
gates on the PCA-unmixed unstained and spectral ribbon plots of the AF
exclusion from the unstained.}
}
\value{
A vector with the indexes of events inside the initial gate.
}
\description{
This function returns a vector with the indexes of events inside the initial
gate on autofluorescence parameters. It proceeds through several steps to
define the gate boundaries and identify density maxima using numerical
search and Voronoi tessellations.
}
