% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate.r
\name{do.gate}
\alias{do.gate}
\title{Do Gate}
\usage{
do.gate(
  gate.data,
  viability.gate,
  large.gate,
  samp,
  scatter.and.channel.label,
  control.type,
  asp
)
}
\arguments{
\item{gate.data}{A data frame containing the gate data.}

\item{viability.gate}{A logical vector indicating the viability gate.}

\item{large.gate}{A logical vector indicating the large gate.}

\item{samp}{A sample identifier.}

\item{scatter.and.channel.label}{A label for scatter and channel.}

\item{control.type}{The type of control used, beads or cells.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A vector with the indexes of events inside the initial gate.
}
\description{
Perform gating on scatter parameters
    Returns a vector with the indexes of events inside the initial gate on
    scatter parameters.

    Proceeds in three steps:
\itemize{
  \item Defines bound by data trimming
  \item Defines region around the target maximum found in bound
  \item Defines gate around target maximum, only in region
}

    Uses numerical search of maxima over estimated densities.
    Uses Voronoi tessellations to improve density estimation around maxima.
}
