% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate.R
\name{do.gate}
\alias{do.gate}
\title{Do Gate}
\usage{
do.gate(
  gate.data,
  viability.gate,
  large.gate,
  samp,
  scatter.and.channel.label,
  control.type,
  asp
)
}
\arguments{
\item{gate.data}{A data frame containing the gate data.}

\item{viability.gate}{A logical vector indicating the viability gate.}

\item{large.gate}{A logical vector indicating the large gate.}

\item{samp}{A sample identifier.}

\item{scatter.and.channel.label}{A label for scatter and channel.}

\item{control.type}{The type of control used, either "beads" or "cells".}

\item{asp}{The AutoSpectral parameter list, prepared using
\code{get.autospectral.param}.}
}
\value{
A vector with the indexes of events inside the initial gate.
}
\description{
Performs gating on scatter parameters and returns a vector with the indexes
of events inside the initial gate.

The gating proceeds in three steps:
\itemize{
\item Defines bounds by data trimming
\item Defines a region around the target maximum found within the bounds
\item Defines a gate around the target maximum, only within that region
}

The method uses numerical search of maxima over estimated densities and
Voronoi tessellations to improve density estimation around maxima.
}
