% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_control.R
\name{downsample.control}
\alias{downsample.control}
\title{Downsample Control Data}
\usage{
downsample.control(
  clean.expr.data,
  samp,
  peak.channels,
  negative.n,
  positive.n,
  verbose = TRUE
)
}
\arguments{
\item{clean.expr.data}{A list containing cleaned expression data for each
sample.}

\item{samp}{The sample identifier.}

\item{peak.channels}{A vector of peak channels for the samples.}

\item{negative.n}{The number of negative events to select.}

\item{positive.n}{The number of positive events to select.}

\item{verbose}{Logical. Default is \code{TRUE}.}
}
\value{
A matrix with the selected expression data.
}
\description{
Downsamples control data by selecting a specified number of
positive and negative events based on peak channel values.
}
