% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_robust_linear_model.r
\name{fit.robust.linear.model}
\alias{fit.robust.linear.model}
\title{Fit Robust Linear Model}
\usage{
fit.robust.linear.model(x.data, y.data, x.name, y.name, asp)
}
\arguments{
\item{x.data}{A vector containing the predictor variable data.}

\item{y.data}{A vector containing the response variable data.}

\item{x.name}{The name of the predictor variable.}

\item{y.name}{The name of the response variable.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.
A list of essential parameters.}
}
\value{
A matrix with the intercept and p-value, and coefficient and p-value.
}
\description{
Returns a matrix by rows, with the intercept and p-value, and
    coefficient and p-value, of a robust linear model fitted to the input data.
    Reverts to a standard linear model in case of no convergence.
}
\details{
Fit Robust Linear Model

This function returns a matrix by rows, with the intercept and p-value,
    and coefficient and p-value, of a robust linear model fitted to the
    input data. It reverts to a standard linear model in case of no convergence.
}
