% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_my_unmix.R
\name{fix.my.unmix}
\alias{fix.my.unmix}
\title{Fix My Unmix}
\usage{
fix.my.unmix(
  spectra,
  unstained.sample,
  fully.stained.sample,
  flow.control,
  asp,
  large.gate = TRUE,
  max.iter = 20,
  downsample = 20000,
  unstained.threshold = 0.99,
  unstained.margin = 1.3,
  convergence.threshold = 0.03,
  biexp = FALSE,
  rs.lambda = 0.1
)
}
\arguments{
\item{spectra}{The spectral matrix, fluorophores x detectors. In this case,
it does not need to be a perfect match for the fluorophores in the fully
stained sample and may, for instance, be pulled from a database for the same
specification of cytometer.}

\item{unstained.sample}{File path and name for a raw unstained sample,
ideally acquired the same day, and ideally matching the autofluorescence of
the fully stained sample.}

\item{fully.stained.sample}{File path and name for a raw fully stained sample.}

\item{flow.control}{The flow.control list.}

\item{asp}{The AutoSpectral parameter list.}

\item{large.gate}{Logical, whether to use a large gate for identifying cells.
Default is \code{TRUE} and will work best for any panels containing any
non-lymphocyte markers.}

\item{max.iter}{Numeric. Limits the number of iterations performed by the
algorithm to avoid overfitting.}

\item{downsample}{Logical/numeric. Controls downsampling to speed up the
algorithm. If \code{FALSE}, no downsampling will be performed. If a numeric,
that number of cells will be used for downsampling. For rare markers,
use larger numbers.}

\item{unstained.threshold}{Numeric between 0 and 1, default \verb{0.99.} The
threshold used to determine positivity in the fully stained sample
(the percentile on the unstained sample in that channel).}

\item{unstained.margin}{Numeric, default \code{1.3}. The fudge factor above the
unstained.threshold.}

\item{convergence.threshold}{Numeric, default \code{0.03}. If the algorithm
reaches this minimal level of spillover error, it will stop.}

\item{biexp}{Logical indicating whether to apply biexponential
transformation, default is \code{FALSE}.}

\item{rs.lambda}{Numeric, default \code{0.1}. Controls how rapidly the spillover
corrections are applied in order to reach convergence.}
}
\value{
A list containing the corrected spillover matrix
(fluorophores x fluorophores), the corrected compensation matrix and the
corrected spectra for unmixing (fluorophores x detectors).
}
\description{
Attempt to automatically fix unmixing errors in fully stained
unmixed samples.
}
