% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_af_sample_plot.R
\name{gate.af.sample.plot}
\alias{gate.af.sample.plot}
\title{Plot Autofluorescence Gates on Samples}
\usage{
gate.af.sample.plot(
  plot.data,
  samp,
  af.boundary.upper,
  asp,
  max.points = 1e+05,
  color.palette = "viridis"
)
}
\arguments{
\item{plot.data}{Matrix containing autofluorescence data points.}

\item{samp}{Sample identifier.}

\item{af.boundary.upper}{Matrix containing upper boundary information.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{max.points}{Number of points to plot (speeds up plotting). Default is
\code{1e5}.}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{viridis}. Options
are the viridis color options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis},
\code{cividis}, \code{rocket}, \code{mako} and \code{turbo}.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots the autofluorescence exclusion gate on the sample(s).
}
