% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gate_sample.r
\name{gate.sample.plot}
\alias{gate.sample.plot}
\title{Plot Pre-defined Gate on Sample}
\usage{
gate.sample.plot(
  samp,
  gate.data,
  gate.marker,
  gate.boundary,
  scatter.and.channel.label,
  control.type,
  asp
)
}
\arguments{
\item{samp}{Sample identifier.}

\item{gate.data}{Matrix containing gate data points.}

\item{gate.marker}{Vector containing gate marker names.}

\item{gate.boundary}{List containing gate boundary information.}

\item{scatter.and.channel.label}{Named vector mapping scatter and channel labels.}

\item{control.type}{Type of control ("beads" or "cells").}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots a pre-defined gate on a sample,
    using ggplot2 and other necessary packages.
}
