% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_sample_plot.r
\name{gate.sample.plot}
\alias{gate.sample.plot}
\title{Plot Pre-defined Gate on Sample}
\usage{
gate.sample.plot(
  samp,
  gate.data,
  gate.marker,
  gate.boundary,
  scatter.and.channel.label,
  control.type,
  asp,
  color.palette = "mako",
  max.points = 50000,
  gate.color = "darkgoldenrod1",
  switch.n = 10000,
  raster.bins = 256L
)
}
\arguments{
\item{samp}{Sample identifier.}

\item{gate.data}{Matrix containing gate data points.}

\item{gate.marker}{Vector containing gate marker names.}

\item{gate.boundary}{List containing gate boundary information.}

\item{scatter.and.channel.label}{Named vector mapping scatter and
channel labels.}

\item{control.type}{Type of control: \code{beads} or \code{cells}. Deprecated.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{color.palette}{Optional character string defining the viridis color
palette to be used for the fluorophore traces. Default is \code{mako}. Use \code{rainbow}
to be similar to FlowJo or SpectroFlo. Other options are the viridis color
options: \code{magma}, \code{inferno}, \code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako}
and \code{turbo}.}

\item{max.points}{Number of points to plot (speeds up plotting). Default is
\code{5e4}.}

\item{gate.color}{Color to plot the gate boundary line, default is \code{darkgoldenrod1}.}

\item{switch.n}{Minimum number of points required to render density contours.
Below this threshold only the rasterised scatter layer is shown. Default is \code{1e4}.}

\item{raster.bins}{Number of pixels on each axis of the rasterised scatter
image. Higher values give finer detail at the cost of a little extra memory.
Default is \code{256L}.}
}
\value{
Saves the plot as a JPEG file in the specified directory.
}
\description{
This function plots a pre-defined gate on a sample, using ggplot2 and other
necessary packages.
}
