% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_af_spectra.R
\name{get.af.spectra}
\alias{get.af.spectra}
\title{Get Autofluorescence Spectra}
\usage{
get.af.spectra(
  unstained.sample,
  asp,
  spectral.channels = flow.control$spectral.channel,
  similarity = 0.99,
  max.spectra = 20,
  min.spectra = 5,
  figures = TRUE,
  plot.dir = NULL,
  plot.title = "Autofluorescence spectra"
)
}
\arguments{
\item{unstained.sample}{Path and file name for a unstained sample FCS file.
The sample type and processing (protocol) method should match the fully
stained samples to which the AF will be applied, ideally.}

\item{asp}{The AutoSpectral parameter list.}

\item{spectral.channels}{The fluorescence detector channels that will be used
to identify the autofluorescence. These should match the channels used for
subsequent unmixing and should normally be all fluorescent area measurements.
Default is \code{flow.control$spectral.channel}.}

\item{similarity}{Numeric between 0 and 1, default is \code{0.99}. The cosine
similarity value that will trigger merging of FlowSOM nodes to create
metaclusters that define the autofluorescence signatures.}

\item{max.spectra}{Maximum number of AF signatures to return. Default is \code{20}.}

\item{min.spectra}{Minimum number of AF signatures to return. Default is \code{5}.}

\item{figures}{Logical, whether to plot the spectral traces and heatmap for
the AF signatures. Default is \code{TRUE}.}

\item{plot.dir}{Directory (folder) where the plots will be saved. Default is
\code{NULL}, which inherits from \code{asp$figure.af.dir}.}

\item{plot.title}{Title for the output spectral plots. Default is
\verb{Autofluorescence spectra}.}
}
\value{
A matrix of autofluorescence spectra.
}
\description{
Extracts autofluorescence spectra from an unstained samples. Intended for use
with \code{unmix.autospectral}. Uses FlowSOM (EmbedSOM) clustering for rapid
identification of cells with similar AF profiles.
}
