% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_af_spectra.R
\name{get.af.spectra}
\alias{get.af.spectra}
\title{Get Autofluorescence Spectra}
\usage{
get.af.spectra(
  unstained.sample,
  spectral.channels = flow.control$spectral.channel,
  threads = asp$worker.process.n,
  n.spectra = 20,
  figures = TRUE,
  plot.dir = asp$figure.af.dir,
  plot.title = "Autofluorescence spectra"
)
}
\arguments{
\item{unstained.sample}{Path and file name for a unstained sample FCS file.
The sample type and processing (protocol) method should match the fully
stained samples to which the AF will be applied, ideally.}

\item{spectral.channels}{The fluorescence detector channels that will be used
to identify the autofluorescence. These should match the channels used for
subsequent unmixing and should normally be all fluorescent area measurements.
Default is \code{flow.control$spectral.channel}}

\item{threads}{Numeric. Number of threads to use for parallel processing in
the creation of the SOM. Default is \code{asp$worker.process.n}.}

\item{n.spectra}{Number of AF signatures to look for. Default is \code{20}.}

\item{figures}{Logical, whether to plot the spectral traces and heatmap for
the AF signatures. Default is \code{TRUE}.}

\item{plot.dir}{Directory (folder) where the plots will be saved. Default is
\code{asp$figure.af.dir}.}

\item{plot.title}{Title for the output spectral plots. Default is
\verb{Autofluorescence spectra}.}
}
\value{
A matrix of autofluorescence spectra.
}
\description{
Extracts autofluorescence spectra from an unstained samples. Intended for use
with \code{unmix.autospectral}. Uses FlowSOM (EmbedSOM) clustering for rapid
identification of cells with similar AF profiles.
}
