% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_af_spectra.R
\name{get.af.spectra}
\alias{get.af.spectra}
\title{Get Autofluorescence Spectra}
\usage{
get.af.spectra(
  unstained.sample,
  asp,
  spectra,
  som.dim = 10,
  figures = TRUE,
  plot.dir = NULL,
  table.dir = NULL,
  title = NULL,
  verbose = TRUE
)
}
\arguments{
\item{unstained.sample}{Path and file name for a unstained sample FCS file.
The sample type and processing (protocol) method should match the fully
stained samples to which the AF will be applied, ideally.}

\item{asp}{The AutoSpectral parameter list.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{som.dim}{Number of x and y dimensions for the SOM. Default is \code{10}.}

\item{figures}{Logical, whether to plot the spectral traces and heatmap for
the AF signatures. Default is \code{TRUE}.}

\item{plot.dir}{Directory (folder) where the plots will be saved. Default is
\code{NULL}, which inherits from \code{asp$figure.af.dir}.}

\item{table.dir}{Directory (folder) where the spectra csv file will be saved.
Default is \code{NULL}, which inherits from \code{asp$table.af.dir}.}

\item{title}{Title for the output spectral plots and csv file. Default is
\verb{Autofluorescence spectra}.}

\item{verbose}{Logical, controls messaging. Default is \code{TRUE}.}
}
\value{
A matrix of autofluorescence spectra.
}
\description{
Extracts autofluorescence spectra from an unstained samples. Intended for use
with \code{unmix.autospectral}. Uses FlowSOM (EmbedSOM) clustering for rapid
identification of cells with similar AF profiles.
}
