% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param_a8.r
\name{get.autospectral.param.a8}
\alias{get.autospectral.param.a8}
\title{Get Autospectral Parameters for DiscoverA8 Cytometer}
\usage{
get.autospectral.param.a8(autosp.param)
}
\arguments{
\item{autosp.param}{A list of initial AutoSpectral parameters.}
}
\value{
A list of AutoSpectral parameters specific to the DiscoverA8
cytometer.
}
\description{
Returns parameters for running a calculation of unmixing with
AutoSpectral, without creating any figures or tables.
}
