% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospectral_param.r
\name{get.autospectral.param}
\alias{get.autospectral.param}
\title{Get Autospectral Parameters}
\usage{
get.autospectral.param(cytometer = "aurora", figures = TRUE)
}
\arguments{
\item{cytometer}{The type of cytometer (default is "aurora").}

\item{figures}{Logical indicating whether to set up directory parameters for
figures and tables (default is TRUE).}
}
\value{
A list of autospectral parameters.
}
\description{
Retrieves autospectral parameters for a specified cytometer.
}
\details{
Get Autospectral Parameters

This function retrieves autospectral parameters for a specified cytometer.
    It sets up directory parameters for figures and tables if required.
}
