% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fluor_variants.R
\name{get.fluor.variants}
\alias{get.fluor.variants}
\title{Get Fluorophore Variants}
\usage{
get.fluor.variants(
  fluor,
  file.name,
  control.dir,
  spectra,
  af.spectra,
  spectral.channel,
  universal.negative,
  control.type,
  raw.thresholds,
  unmixed.thresholds,
  flow.channel,
  som.dim,
  n.cells,
  asp,
  verbose,
  output.dir,
  sim.threshold,
  figures
)
}
\arguments{
\item{fluor}{The name of the fluorophore.}

\item{file.name}{A named vector of file names for the samples.}

\item{control.dir}{The directory containing the control files.}

\item{spectra}{A matrix containing the spectral data. Fluorophores in rows,
detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{spectral.channel}{A vector of spectral channels.}

\item{universal.negative}{A named vector of unstained negative samples, with
names corresponding to the fluorophores.}

\item{control.type}{Character, either "beads" or "cells". Determines the type
of control sample being used and the subsequent processing steps.}

\item{raw.thresholds}{A named vector of numerical values corresponding to
the threshold for positivity in each raw detector channel. Determined by the
99.5th percentile on the unstained sample, typically.}

\item{unmixed.thresholds}{A named vector of numerical values corresponding to
the threshold for positivity in each unmixed channel. Determined by the
99.5th percentile on the unstained sample, typically after single-cell AF
unmixing.}

\item{flow.channel}{A named vector of peak raw channels, one per fluorophore.}

\item{som.dim}{Numeric, default \code{10}. Number of x and y dimensions to use in
the SOM for clustering the spectral variation.}

\item{n.cells}{Numeric, default \code{2000}. Number of cells to use for defining
the variation in spectra. Up to \code{n.cells} cells will be selected as positive
events in the peak channel for each fluorophore, above the \code{pos.quantile} in
the unstained sample.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}

\item{output.dir}{File path to whether the figures and .rds data file will be
saved. Default is \code{NULL}, in which case \code{asp$variant.dir} will be used.}

\item{sim.threshold}{Numeric, default \code{0.98}. Threshold for cosine similarity-
based exclusion of spectral variants. Any variant less than \code{sim.threshold}
by \code{cosine.similarity} from the optimized spectrum for that fluorophore (from
\code{spectra}) will be excluded from output. This helps to exclude autofluorescence
contamination.}

\item{figures}{Logical, controls whether the variation in spectra for each
fluorophore is plotted in \code{output.dir}. Default is \code{TRUE}.}
}
\value{
A matrix with the flow expression data.
}
\description{
Assesses variation in the spectral signature of a single-stained flow
cytometry control sample. Uses SOM-based clustering on the brightest positive
events in the file.
}
