% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fluorophore_spectra.r
\name{get.fluorophore.spectra}
\alias{get.fluorophore.spectra}
\title{Get Fluorophore Spectra}
\usage{
get.fluorophore.spectra(
  flow.control,
  asp,
  use.clean.expr = FALSE,
  biexp = FALSE,
  af.spectra = NULL,
  plot.prefix = "Initial"
)
}
\arguments{
\item{flow.control}{A list containing flow cytometry control data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{use.clean.expr}{Logical indicating whether to use cleaned expression
data, default is \code{FALSE}}

\item{biexp}{Logical indicating whether to apply biexponential
transformation, default is \code{FALSE}}

\item{af.spectra}{Optional autofluorescence spectra to include.}

\item{plot.prefix}{Optional prefix for plot titles, default is \code{Initial}.}
}
\value{
A matrix with the fluorophore spectra.
}
\description{
Retrieves the fluorophore spectra for flow cytometry data,
optionally using cleaned expression data and biexponential transformation.
}
\details{
Get Fluorophore Spectra

This function retrieves the fluorophore spectra for flow cytometry data,
optionally using cleaned expression data and biexponential transformation.
It also plots and saves the spectra, and performs cosine similarity QC for
controls.
}
