% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fluorophore_spectra.r
\name{get.fluorophore.spectra}
\alias{get.fluorophore.spectra}
\title{Get Fluorophore Spectra}
\usage{
get.fluorophore.spectra(
  flow.control,
  asp,
  use.clean.expr = TRUE,
  af.spectra = NULL,
  title = NULL,
  figures = TRUE
)
}
\arguments{
\item{flow.control}{A list containing flow cytometry control data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{use.clean.expr}{Logical indicating whether to use cleaned expression
data, default is \code{TRUE}}

\item{af.spectra}{Optional autofluorescence spectra to include.}

\item{title}{Optional prefix for plot titles, default is \code{NULL}, which gives
"Initial" when \code{use.clean.expr} is \code{FALSE} and "Clean" when \code{use.clean.expr}
is \code{TRUE}.}

\item{figures}{Logical, default is \code{TRUE}. Whether to produce plots of the
fluorophore spectra and cosine similarity.}
}
\value{
A matrix with the fluorophore spectra.
}
\description{
This function retrieves the fluorophore spectra for flow cytometry data,
optionally using cleaned expression data.
It also plots and saves the spectra, and performs cosine similarity QC for
controls.
}
