% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectral_variants.R
\name{get.spectral.variants}
\alias{get.spectral.variants}
\title{Get Spectral Variations for Fluorophores}
\usage{
get.spectral.variants(
  control.dir,
  control.def.file,
  asp,
  spectra,
  af.spectra,
  n.cells = 2000,
  som.dim = 10,
  figures = TRUE,
  output.dir = NULL,
  parallel = FALSE,
  verbose = TRUE,
  threads = NULL,
  refine = TRUE,
  problem.quantile = 0.95,
  ...
)
}
\arguments{
\item{control.dir}{File path to the single-stained control FCS files.}

\item{control.def.file}{CSV file defining the single-color control file names,
fluorophores they represent, marker names, peak channels, and gating requirements.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{spectra}{A matrix containing the spectral data. Fluorophores in rows,
detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{n.cells}{Numeric, default \code{2000}. Number of cells to use for defining
the variation in spectra. Up to \code{n.cells} cells will be selected as positive
events in the peak channel for each fluorophore, above the 99.5th percentile
level in the unstained sample.}

\item{som.dim}{Numeric, default \code{10}. Number of x and y dimensions to use in
the SOM for clustering the spectral variation. The number of spectra returned
for each fluorophore will increase with the quadratic of \code{som.dim}, so for 10,
you will get up to 100 variants. Somewhere between 4 and 7 appears to be
sufficient, but with the pruning of variants implemented in
\code{unmix.autospectral()} in v1.0.0, this is less important.}

\item{figures}{Logical, controls whether the variation in spectra for each
fluorophore is plotted in \code{output.dir}. Default is \code{TRUE}.}

\item{output.dir}{File path to whether the figures and .rds data file will be
saved. Default is \code{NULL}, in which case \code{asp$variant.dir} will be used.}

\item{parallel}{Logical, default is \code{FALSE}, in which case sequential processing
will be used. The new parallel processing should always be faster.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}.}

\item{refine}{Logical, default is \code{TRUE}. Controls whether to perform a second
round of variation measurement on "problem cells", which are those with the
highest spillover, as defined by \code{problem.quantile}.}

\item{problem.quantile}{Numeric, default \code{0.95}. The quantile for determining
which cells will be considered "problematic" after unmixing with per-cell AF
extraction. Cells in the \code{problem.quantile} or above with respect to total
signal in the fluorophore (non-AF) channels after per-cell AF extraction will
be used to determine additional autofluorescence spectra, using a second round
of clustering and modulation of the previously selected autofluroescence
spectra. A value of \code{0.95} means the top 5\% of cells, those farthest from zero,
will be selected for further investigation.}

\item{...}{Ignored. Previously used for deprecated arguments such as
\code{pos.quantile} and \code{sim.threshold}, which are now fixed internally and no
longer user-settable.}
}
\value{
A vector with the indexes of events inside the initial gate.
}
\description{
This function cycles through all the fluorophores defined in \code{control.def.file},
identifying variations in spectral profiles. It does this by performing SOM
clustering on the positive events in the cleaned control data. The output is
saved as an .rds file, and figures summarizing the variation are saved, if
desired. Note that the .rds file contains all the needed information for
downstream processing (per-cell unmixing), so you can just load that using
the \code{readRDS} function) rather than re-running this process.
}
