% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_sequential.R
\name{lapply.sequential}
\alias{lapply.sequential}
\title{Sequential lapply}
\usage{
lapply.sequential(X, FUN, ..., future.seed = NULL)
}
\arguments{
\item{X}{A vector (atomic or list) or an expression object. Other objects
(including classed objects) will be coerced by `as.list`.}

\item{FUN}{The function to be applied to each element of `X`.}

\item{...}{Optional arguments to `FUN`.}

\item{future.seed}{Ignored in this function, included for compatibility with
`future_lapply`.}
}
\value{
A list of the same length as `X` and named by `X`.
}
\description{
This function allows switching between `lapply` and `future_lapply`
    without `future.seed` issues by using `lapply` sequentially.
}
