% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fluorophores.R
\name{match.fluorophores}
\alias{match.fluorophores}
\title{Match Fluorophores}
\usage{
match.fluorophores(control.filenames, fluorophore.database)
}
\arguments{
\item{control.filenames}{Vector of control filenames.}

\item{fluorophore.database}{Data frame containing fluorophore information.}
}
\value{
A named vector of matched fluorophores for each control filename.
}
\description{
This function matches control filenames to fluorophores in the fluorophore
database, including synonyms, and returns the matched fluorophores based on
the longest character match to ensure specificity.
}
