% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_markers.R
\name{match.markers}
\alias{match.markers}
\title{Match Markers}
\usage{
match.markers(control.filenames, marker.database)
}
\arguments{
\item{control.filenames}{Vector of control filenames.}

\item{marker.database}{Data frame containing marker information.}
}
\value{
A named vector of matched markers for each control filename.
}
\description{
This function matches control filenames to markers in the marker database,
including synonyms, and returns the matched markers
}
