% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_unmix.R
\name{optimize.unmix}
\alias{optimize.unmix}
\title{Optimize Spectral Unmixing}
\usage{
optimize.unmix(
  raw.data,
  unmixed,
  spectra,
  pos.thresholds,
  optimize.fluors,
  variants,
  delta.list,
  delta.norms,
  fluorophores,
  asp,
  k = 10L,
  nthreads = 1L,
  parallel = TRUE
)
}
\arguments{
\item{raw.data}{Numeric matrix (cells x detectors)}

\item{unmixed}{Numeric matrix (cells x fluors)}

\item{spectra}{Numeric matrix (fluors x detectors)}

\item{pos.thresholds}{Numeric vector (n fluors)}

\item{optimize.fluors}{Character vector of fluorophores present in variants}

\item{variants}{List of variant matrices per fluorophore}

\item{delta.list}{List of delta matrices per fluorophore}

\item{delta.norms}{List of delta norms per fluorophore}

\item{fluorophores}{Character vector of fluorophore names}

\item{asp}{The AutoSpectral parameter list.}

\item{k}{Integer, number of variants to test}

\item{nthreads}{Integer, number of threads}

\item{parallel}{Logical, whether to use parallel processing}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Parallel backend for per-cell spectral optimization in R.
}
