% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_af_spectra.R
\name{qc.af.spectra}
\alias{qc.af.spectra}
\title{QC Autofluorescence Spectra}
\usage{
qc.af.spectra(
  af.spectra,
  spectra,
  output.dir = "./figure_autofluorescence",
  remove = TRUE,
  pass = 1
)
}
\arguments{
\item{af.spectra}{Matrix or dataframe containing autofluorescence spectral
signatures. This should be in format AFn x detectors. Row names will be used
as the names.}

\item{spectra}{Matrix or dataframe containing spectral data. This
should be in format fluorophores x detectors. Row names will be used as the
fluorophore names. Column names will be used as the detectors (channels).}

\item{output.dir}{Directory where the files will be saved.
Default is \code{./figure_autofluorescence}.}

\item{remove}{Logical. When \code{TRUE}, identified highly similar AF spectra will
be removed, returning only AF spectra that at least 0.995 or less from any
fluorophore spectrum.}

\item{pass}{Numeric, default \code{1}. Counter to separate multiple passes of AF
extraction, such as occur when \code{refine=TRUE} in \code{get.af.spectra()}.}
}
\value{
None. Plots are saved to a PDF in \code{plot.dir}.
}
\description{
This function performs a comparison between the user's fluorophore spectra,
from \code{spectra}, and identified autofluorescence spectra to check for any
contamination. Contamination may occur if the unstained sample used to generate
the AF spectra was not actually unstained, contained fluorescent protein
reporter constructs, or was contaminated during acquisition, for example by
crossover from nearby wells on a plate. Quality control is performed using
cosine similarity.
}
\seealso{
\itemize{
\item \code{\link[=af.qc.plot]{af.qc.plot()}}
\item \code{\link[=get.af.spectra]{get.af.spectra()}}
}
}
