% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read.spectra}
\alias{read.spectra}
\title{Read In Saved Spectra}
\usage{
read.spectra(
  spectra.file,
  spectra.dir = "./table_spectra",
  remove.af = FALSE,
  af.param = "AF"
)
}
\arguments{
\item{spectra.file}{File name for the spectra CSV file to be read.}

\item{spectra.dir}{File path to the folder containing \code{spectra.file}. Default
is \code{table_spectra}, where \code{AutoSpectral} saves the spectra files.}

\item{remove.af}{Logical, default is \code{FALSE}. If \code{TRUE}, returns the spectral
matrix without the default autofluorescence spectrum.}

\item{af.param}{Name of the autofluorescence parameter. Default is \code{AF}. Note
that any fluorophores can be removed from the matrix by supplying a character
vector, e.g., \code{c("BUV395", "PE")}, if desired.}
}
\value{
A matrix containing the fluorophore spectra (fluorophore x detectors).
}
\description{
Reads in CSV files created by \code{AutoSpectral} or in the same format
(fluorophores in rows, detectors in columns, first row is detector names,
first column contains fluorophore names).
}
