% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_flow_control.R
\name{reload.flow.control}
\alias{reload.flow.control}
\title{Reload Flow Control Information}
\usage{
reload.flow.control(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{file path to the single stained control FCS files}

\item{control.def.file}{csv file defining the single color control file
names, fluorophores they represent, marker names, peak channels and
gating requirements.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A list containing the reloaded flow control information.
}
\description{
This function reloads essential information from control files to permit
rapid unmixing at a later date, without recalculating spectra or gates.
}
