% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_af.R
\name{remove.af}
\alias{remove.af}
\title{Remove Autofluorescence Contamination}
\usage{
remove.af(
  clean.expr.data,
  samp,
  af.artefact,
  spectral.channel,
  universal.negative,
  asp
)
}
\arguments{
\item{clean.expr.data}{List containing cleaned expression data.}

\item{samp}{Sample identifier.}

\item{af.artefact}{List containing autofluorescence artefacts.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{universal.negative}{Named vector mapping samples to their
universal negatives.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A matrix containing the expression data with autofluorescent events
removed for the sample.
}
\description{
This function removes autofluorescence contamination from a
sample, using the specified parameters and settings.
}
