% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_af_removal.R
\name{run.af.removal}
\alias{run.af.removal}
\title{Run Autofluorescence Removal}
\usage{
run.af.removal(
  clean.expr,
  af.removal.sample,
  neg.artefacts.list,
  spectral.channel,
  universal.negative,
  asp
)
}
\arguments{
\item{clean.expr}{List containing cleaned expression data.}

\item{af.removal.sample}{Vector of sample names for which autofluorescence
removal is to be performed.}

\item{neg.artefacts.list}{List of negative artefacts.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{universal.negative}{Name of the universal negative control.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}
}
\value{
A list containing the expression data with autofluorescence removed for each sample.
}
\description{
This function runs the autofluorescence removal process on a
    list of samples, using the specified parameters and settings.
}
