% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_af_removal.R
\name{run.af.removal}
\alias{run.af.removal}
\title{Run Autofluorescence Removal}
\usage{
run.af.removal(
  clean.expr,
  af.removal.sample,
  spectral.channel,
  peak.channel,
  universal.negative,
  asp,
  scatter.param,
  negative.n = 500,
  positive.n = 1000,
  scatter.match = TRUE,
  intermediate.figures = FALSE,
  main.figures = TRUE,
  parallel = FALSE,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{clean.expr}{List containing cleaned expression data.}

\item{af.removal.sample}{Vector of sample names for which autofluorescence
removal is to be performed.}

\item{spectral.channel}{Vector of spectral channel names.}

\item{peak.channel}{Vector of peak detection channels for fluorophores.}

\item{universal.negative}{Name of the universal negative control.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{scatter.param}{Vector of scatter parameters.}

\item{negative.n}{Integer. Number of events to include in the downsampled
negative population. Default is \code{500}.}

\item{positive.n}{Integer. Number of events to include in the downsampled
positive population. Default is \code{1000}.}

\item{scatter.match}{Logical, default is \code{TRUE}. Whether to select negative
events based on scatter profiles matching the positive events. Defines a
region of FSC and SSC based on the distribution of selected positive events.}

\item{intermediate.figures}{Logical, if \code{TRUE} returns additional figures to
show the inner workings of the cleaning, including definition of low-AF cell
gates on the PCA-unmixed unstained and spectral ribbon plots of the AF
exclusion from the unstained.}

\item{main.figures}{Logical, if \code{TRUE} creates the main figures to show the
impact of intrusive autofluorescent event removal and scatter-matching for
the negatives.}

\item{parallel}{Logical, default is \code{FALSE}, in which case parallel processing
will not be used. Set to \code{TRUE} to run in parallel.}

\item{threads}{Number of cores to use for parallel processing, default is \code{1}.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress messages.}
}
\value{
A list containing the expression data with autofluorescent events
removed for each sample.
}
\description{
This function runs the autofluorescence removal process on a list of samples,
using the specified parameters and settings.
}
