% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_downsample.R
\name{run.downsample}
\alias{run.downsample}
\title{Run Downsample}
\usage{
run.downsample(
  clean.expr.data,
  downsample.sample,
  peak.channels,
  negative.n,
  positive.n,
  asp
)
}
\arguments{
\item{clean.expr.data}{List containing cleaned expression data.}

\item{downsample.sample}{Vector of sample names to be downsampled.}

\item{peak.channels}{Named vector mapping samples to their peak channels.}

\item{negative.n}{Number of negative events to select.}

\item{positive.n}{Number of positive events to select.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A list containing the downsampled expression data for each sample.
}
\description{
This function runs the downsampling process on a list of
samples, using the specified peak channels and parameters.
}
