% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_peacoqc.R
\name{run.peacoQC}
\alias{run.peacoQC}
\title{Run PeacoQC}
\usage{
run.peacoQC(expr.data, spectral.channel, all.channels, asp)
}
\arguments{
\item{expr.data}{A list containing the expression data for each sample.}

\item{spectral.channel}{A character vector specifying the spectral channels.}

\item{all.channels}{A character vector specifying all channels.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}
}
\value{
A list containing the cleaned expression data for each sample.
}
\description{
This function runs PeacoQC to remove flow fluctuation errors from expression
data using parallel processing if specified.
}
