% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_reference_plot.R
\name{spectral.reference.plot}
\alias{spectral.reference.plot}
\title{Spectral Reference Plot}
\usage{
spectral.reference.plot(
  spectra,
  asp,
  qc.threshold = 0.98,
  experiment.control.color = "black",
  library.reference.color = "blue",
  experiment.line.type = "solid",
  library.line.type = "dotted",
  pass.color = "darkgreen",
  fail.color = "red",
  linewidth = 1,
  plot.dir = "./figure_spectra",
  filename = "spectral_qc_report.pdf"
)
}
\arguments{
\item{spectra}{Matrix or dataframe containing spectral data. This
should be in format fluorophores x detectors. Row names will be used as the
fluorophore names. Column names will be used as the detectors (channels).}

\item{asp}{The AutoSpectral parameter list. Used to determine which cytometer
produced the data.}

\item{qc.threshold}{Numeric, default \code{0.98}. The similarity value to trigger
a QC failure warning.}

\item{experiment.control.color}{Color for the line representing the user's
fluorophore spectrum from the single-stained reference control. Default is
\code{black}.}

\item{library.reference.color}{Color for the line representing the library
reference standard fluorophore spectrum. Default is \code{blue}.}

\item{experiment.line.type}{Line style for the line representing the user's
fluorophore spectrum from the single-stained reference control. Default is
\code{solid}.}

\item{library.line.type}{Line style for the line representing the library
reference standard fluorophore spectrum. Default is \code{dotted}.}

\item{pass.color}{Color to label similarity values above the \code{qc.threshold},
i.e., fluorophores passing QC. Default is \code{darkgreen}.}

\item{fail.color}{Color to label similarity values below the \code{qc.threshold},
i.e., fluorophores failing QC. Default is \code{red}.}

\item{linewidth}{Width of the line for the spectral traces. Default is \code{1}.}

\item{plot.dir}{Directory where the files will be saved.
Default is \code{./figure_spectra}.}

\item{filename}{Name for the output PDF file. Default is \code{spectral_qc_report.pdf}.}
}
\value{
None. Plots are saved to a PDF in \code{plot.dir}.
}
\description{
This function plots a comparison between the user's fluorophore spectrum,
from \code{spectra}, and a known reference spectrum (if available) for the same
fluorophore. Quality control is performed using cosine similarity between the
two spectral profiles.
}
