% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{spectral.trace}
\alias{spectral.trace}
\title{Plot Fluorophore Spectra Traces}
\usage{
spectral.trace(
  spectral.matrix,
  flow.control,
  asp,
  plot.title = "Fluorophore Spectra",
  plot.dir,
  split.lasers = TRUE
)
}
\arguments{
\item{spectral.matrix}{Matrix containing spectral data.}

\item{flow.control}{List containing flow control information, including
spectral channels.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{plot.title}{Title for the plot.}

\item{plot.dir}{Directory to save the plot files.}

\item{split.lasers}{Logical indicating whether to split the plot by
excitation lasers.}
}
\value{
Saves the plot(s) as JPEG files in the specified directory.
}
\description{
This function plots the fluorophore spectra, optionally splitting
    by excitation lasers, and saves the plots as JPEG files.
}
