% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_autospectral.R
\name{unmix.autospectral}
\alias{unmix.autospectral}
\title{Unmix AutoSpectral}
\usage{
unmix.autospectral(raw.data, spectra, af.spectra, weighted = FALSE)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm. Default is \code{FALSE} and will use OLS.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Unmix using the AutoSpectral method to extract autofluorescence at the
single cell level.
}
