% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_autospectral.R
\name{unmix.autospectral}
\alias{unmix.autospectral}
\title{Unmix AutoSpectral}
\usage{
unmix.autospectral(
  raw.data,
  spectra,
  af.spectra,
  asp,
  spectra.variants = NULL,
  use.dist0 = TRUE,
  verbose = TRUE,
  speed = c("slow", "medium", "fast"),
  parallel = TRUE,
  threads = NULL,
  n.variants = NULL,
  ...
)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{asp}{The AutoSpectral parameter list.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}. Default is \code{NULL}.}

\item{use.dist0}{Logical, controls whether the selection of the optimal AF
signature for each cell is determined by which unmixing brings the fluorophore
signals closest to 0 (\code{use.dist0} = \code{TRUE}) or by which unmixing minimizes the
per-cell residual (\code{use.dist0} = \code{FALSE}). Default is \code{TRUE}.}

\item{verbose}{Logical, default \code{TRUE}. Whether to send messages to the console.}

\item{speed}{Selector for the precision-speed trade-off in AutoSpectral per-cell
fluorophore optimization. Options are \code{slow}, \code{medium} and \code{fast}. From v1.0.0,
this controls the number of variants tested per cell (and per fluorophore).
More variants takes longer, but gives better resolution in the unmixed data.
When \code{speed = fast}, as single variant will be tested; for \code{medium}, three
will be tested and for \code{slow}, 10 variants will be tested. From AutoSpectral
v1.0.0, \code{slow} is the default and is available in the pure R version.
Installation of \code{AutoSpectralRcpp} is strongly encouraged for speed, though.}

\item{parallel}{Logical, default is \code{TRUE}. The new parallel processing
should always be faster.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}.}

\item{n.variants}{Number of variants to test per cell. Allows explicit control
over the number used, as opposed to \code{speed}, which selects from pre-defined
choices. Providing a numeric value to \code{n.variants} will override \code{speed},
allowing up to \code{n.variants} (or the max available) variants to be tested. The
default is \code{NULL}, in which case \code{n.variants} will be ignored.}

\item{...}{Ignored. Previously used for deprecated arguments such as
\code{calculate.error}.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Unmix using the AutoSpectral method to extract autofluorescence and optimize
fluorophore signatures at the single cell level.
}
