% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_autospectral.R
\name{unmix.autospectral}
\alias{unmix.autospectral}
\title{Unmix AutoSpectral}
\usage{
unmix.autospectral(
  raw.data,
  spectra,
  af.spectra,
  asp,
  spectra.variants = NULL,
  weighted = FALSE,
  weights = NULL,
  use.dist0 = TRUE,
  verbose = TRUE,
  parallel = TRUE,
  threads = NULL,
  ...
)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{asp}{The AutoSpectral parameter list.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}. Default is \code{NULL}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm. Default is \code{FALSE} and will use OLS.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{NULL}, in which case weighting will be done by
channel means (Poisson variance). Only used if \code{weighted}.}

\item{use.dist0}{Logical, controls whether the selection of the optimal AF
signature for each cell is determined by which unmixing brings the fluorophore
signals closest to 0 (\code{use.dist0} = \code{TRUE}) or by which unmixing minimizes the
per-cell residual (\code{use.dist0} = \code{FALSE}). Default is \code{TRUE}.}

\item{verbose}{Logical, whether to send messages to the console.
Default is \code{TRUE}.}

\item{parallel}{Logical, default is \code{FALSE}, in which case sequential processing
will be used. The new parallel processing should always be faster.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}.}

\item{...}{Ignored. Previously used for deprecated arguments such as
\code{calculate.error}.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Unmix using the AutoSpectral method to extract autofluorescence and optimize
fluorophore signatures at the single cell level.
}
