% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_fcs.R
\name{unmix.fcs}
\alias{unmix.fcs}
\title{Unmix FCS Data}
\usage{
unmix.fcs(
  fcs.file,
  spectra,
  asp,
  flow.control,
  method = "ols",
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  allow.negative = TRUE
)
}
\arguments{
\item{fcs.file}{A character string specifying the path to the FCS file.}

\item{spectra}{A matrix containing the spectral data.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
Options are "ols", "wls", and "poisson". Default is "ols".}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file. Default is NULL.}

\item{file.suffix}{A character string to append to the output file name.
Default is NULL.}

\item{include.raw}{A logical value indicating whether to include raw expression
data in the output. Default is FALSE.}

\item{include.imaging}{A logical value indicating whether to include imaging
parameters in the output. Default is FALSE.}

\item{allow.negative}{A logical value indicating whether to allow negative
values in the unmixing process. Default is TRUE.}
}
\value{
None. The function writes the unmixed FCS data to a file.
}
\description{
This function performs spectral unmixing on FCS data using
    various methods.
}
