% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_fcs.R
\name{unmix.fcs}
\alias{unmix.fcs}
\title{Unmix FCS Data}
\usage{
unmix.fcs(
  fcs.file,
  spectra,
  asp,
  flow.control,
  method = "Automatic",
  weighted = FALSE,
  weights = NULL,
  af.spectra = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  calculate.error = TRUE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5
)
}
\arguments{
\item{fcs.file}{A character string specifying the path to the FCS file.}

\item{spectra}{A matrix containing the spectral data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
The default is \code{Automatic}, which uses \code{AutoSpectral} for AF extraction if
af.spectra are provided and automatically selects \code{OLS} or \code{WLS} depending
on which is normal for the given cytometer in \code{asp$cytometer}. This means
that files from the ID7000, A8 and S8 will be unmixed using \code{WLS} while
others will be unmixed with \code{OLS}. Any option can be set manually.
Manual options are \code{OLS}, \code{WLS}, \code{AutoSpectral}, \code{Poisson} and \code{FastPoisson}.
Default is \code{OLS}. \code{FastPoisson} requires installation of \code{AutoSpectralRcpp}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm in AutoSpectral unmixing.
Default is \code{FALSE} and will use OLS.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{NULL}, in which case weighting will be done by
channel means (Poisson variance). Only used for \code{WLS}.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}. Required for \code{AutoSpectral} unmixing. Default is
\code{NULL} and will thus provoke failure if no spectra are provided and
\code{AutoSpectral} is selected.}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file. Default is \code{NULL}.}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}.}

\item{include.raw}{A logical value indicating whether to include raw
expression data in the written FCS file. Default is \code{FALSE}.}

\item{include.imaging}{A logical value indicating whether to include imaging
parameters in the written FCS file. Default is \code{FALSE}.}

\item{calculate.error}{Logical, whether to calculate the RMSE unmixing model
accuracy and include it as a keyword in the FCS file.}

\item{divergence.threshold}{Numeric. Used for \code{FastPoisson} only.
Threshold to trigger reversion towards WLS unmixing when Poisson result
diverges for a given point.}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are \code{NonNeg} (non-negativity will be enforced), \code{WLS} (revert
to WLS intial unmix) or \code{Balance} (WLS and NonNeg will be averaged).
Default is \code{Balance}}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells.
Used for \code{Balance} option under \code{divergence.handling}. Default is \code{0.5}.}
}
\value{
None. The function writes the unmixed FCS data to a file.
}
\description{
This function performs spectral unmixing on FCS data using various methods.
}
