% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_fcs.R
\name{unmix.fcs}
\alias{unmix.fcs}
\title{Unmix FCS Data}
\usage{
unmix.fcs(
  fcs.file,
  spectra,
  asp,
  flow.control,
  method = "OLS",
  weights = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5
)
}
\arguments{
\item{fcs.file}{A character string specifying the path to the FCS file.}

\item{spectra}{A matrix containing the spectral data.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
Options are "OLS", "WLS", "Poisson" and "FastPoisson". Default is "OLS".
"FastPoisson" requires installation of AutoSpectralRcpp.}

\item{weights}{Optional numeric vector of weights (one per fluorescent detector).
Default is NULL, in which case weighting will be done by channel means.
Only used for WLS}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file. Default is NULL.}

\item{file.suffix}{A character string to append to the output file name.
Default is NULL.}

\item{include.raw}{A logical value indicating whether to include raw expression
data in the output. Default is FALSE.}

\item{include.imaging}{A logical value indicating whether to include imaging
parameters in the output. Default is FALSE.}

\item{divergence.threshold}{Numeric. Used for FastPoisson only. Threshold to
trigger reversion towards WLS unmixing when Poisson result diverge.}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are "NonNeg" (non-negativity will be enforced), "WLS" (revert
to WLS intial unmix) or "Balance" (WLS and NonNeg will be averaged).
Default is "Balance".}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells. Used
for "Balance" option under divergence.handling. Default is 0.5.}
}
\value{
None. The function writes the unmixed FCS data to a file.
}
\description{
This function performs spectral unmixing on FCS data using
    various methods.
}
