% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_fcs.R
\name{unmix.fcs}
\alias{unmix.fcs}
\title{Unmix FCS Data}
\usage{
unmix.fcs(
  fcs.file,
  spectra,
  asp,
  flow.control,
  method = "AutoSpectral",
  weighted = FALSE,
  weights = NULL,
  af.spectra = NULL,
  spectra.variants = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  use.dist0 = TRUE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5,
  speed = "slow",
  parallel = TRUE,
  threads = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fcs.file}{A character string specifying the path to the FCS file.}

\item{spectra}{A matrix containing the spectral data. Fluorophores in rows,
detectors in columns.}

\item{asp}{The AutoSpectral parameter list. Prepare using
\code{get.autospectral.param}.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use. The
default as of version 1.0.0 is now \code{AutoSpectral} to avoid confusion. To use
AutoSpectral unmixing, you must provide at least \code{af.spectra} to perform
autofluorescence extraction (on a per-cell basis). To also optimize
fluorophore spectra, provide \code{spectra.variants}. To perform other types of
unmixing, select from the options: \code{OLS}, \code{WLS}, \code{Poisson} or \code{FastPoisson}.
\code{FastPoisson} requires installation of \code{AutoSpectralRcpp}.There is also
\code{Automatic}, which switches depending on the inputs provided: it uses
\code{AutoSpectral} for AF extraction if af.spectra are provided, and automatically
selects \code{OLS} or \code{WLS} depending on which is normal for the given cytometer
in \code{asp$cytometer}. This means that files from the ID7000, A8 and S8 will be
unmixed using \code{WLS} while others will be unmixed with \code{OLS}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm in AutoSpectral unmixing.
Default is \code{FALSE} and will use OLS.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{NULL}, in which case weighting will be done by
channel means (Poisson variance). Only used for \code{WLS}.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}. Required for \code{AutoSpectral} unmixing. Default is
\code{NULL} and will thus provoke failure if no spectra are provided and
\code{AutoSpectral} is selected.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}. Default is \code{NULL}. Used for
AutoSpectral unmixing. Required for per-cell fluorophore optimization.}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file. Default is \code{NULL}, which will use \code{./AutoSpectral_unmixed}.}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}.}

\item{include.raw}{A logical value indicating whether to include raw
expression data in the written FCS file. Default is \code{FALSE} to provide smaller
output files.}

\item{include.imaging}{A logical value indicating whether to include imaging
parameters in the written FCS file. Default is \code{FALSE} to provide smaller
output files.}

\item{use.dist0}{Logical, controls whether the selection of the optimal AF
signature for each cell is determined by which unmixing brings the fluorophore
signals closest to 0 (\code{use.dist0} = \code{TRUE}) or by which unmixing minimizes the
per-cell residual (\code{use.dist0} = \code{FALSE}). Default is \code{TRUE}. Used for
AutoSpectral unmixing. The minimization of fluorophore signals can be thought
of as a "worst-case" scenario, but it provides more accurate assignments,
particularly with large panels.}

\item{divergence.threshold}{Numeric. Used for \code{FastPoisson} only.
Threshold to trigger reversion towards WLS unmixing when Poisson result
diverges for a given point. To be deprecated.}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are \code{NonNeg} (non-negativity will be enforced), \code{WLS} (revert
to WLS initial unmix) or \code{Balance} (WLS and NonNeg will be averaged).
Default is \code{Balance}. To be deprecated.}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells.
Used for \code{Balance} option under \code{divergence.handling}. Default is \code{0.5}.
To be deprecated.}

\item{speed}{Selector for the precision-speed trade-off in AutoSpectral per-cell
fluorophore optimization. Options are \code{fast}, \code{medium} and \code{slow}, with the
default being \code{slow}. As of version 1.0.0, the backend for how this works
has changed. Spectral variants and AF signatures are now pre-screened per cell
to identify likely candidates, so brute force testing of all variants is no
longer required. So, \code{speed} controls the number of variants to be tested per
cell, with \code{fast} testing a single variant, \code{medium} testing 3 variants, and
\code{slow} testing 10 variants. While this is now implemented in pure R in
\code{AutoSpectral}, installation of \code{AutoSpectralRcpp} is strongly encouraged for
faster processing.}

\item{parallel}{Logical, default is \code{TRUE}, which enables parallel processing
for per-cell unmixing methods.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}. If working on a computing cluster, try \code{parallelly::availableCores()}.}

\item{verbose}{Logical, controls messaging. Default is \code{TRUE}. Set to \code{FALSE}
to have it shut up.}

\item{...}{Ignored. Previously used for deprecated arguments such as
\code{calculate.error}.}
}
\value{
None. The function writes the unmixed FCS data to a file.
}
\description{
This function performs spectral unmixing on FCS data using various methods.
}
