% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_folder.R
\name{unmix.folder}
\alias{unmix.folder}
\title{Unmix All FCS Files in a Directory}
\usage{
unmix.folder(
  fcs.dir,
  raw.channels,
  spectra,
  asp,
  method = "ols",
  output.dir = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  allow.negative = TRUE
)
}
\arguments{
\item{fcs.dir}{Directory containing FCS files to be unmixed.}

\item{raw.channels}{Vector of raw channel names.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{method}{Method to be used for unmixing (default is "ols").}

\item{output.dir}{Directory to save the unmixed FCS files (default is asp$unmixed.fcs.dir).}

\item{include.raw}{Logical indicating whether to include raw data in the output.}

\item{include.imaging}{Logical indicating whether to include imaging data in the output.}

\item{allow.negative}{Logical indicating whether to allow negative coefficients.}
}
\value{
None. Saves the unmixed FCS files to the specified output directory.
}
\description{
This function unmixes all FCS files in a specified directory
    using the provided spectra and method, and saves the unmixed data to an
    output directory.
}
