% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_folder.R
\name{unmix.folder}
\alias{unmix.folder}
\title{Unmix All FCS Files in a Directory}
\usage{
unmix.folder(
  fcs.dir,
  spectra,
  asp,
  flow.control,
  method = "OLS",
  weights = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5
)
}
\arguments{
\item{fcs.dir}{Directory containing FCS files to be unmixed.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list. Prepare using get.autospectral.param.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
Options are "OLS", "WLS", "Poisson" and "FastPoisson". Default is "OLS".
"FastPoisson" requires installation of AutoSpectralRcpp.}

\item{weights}{Optional numeric vector of weights (one per fluorescent detector).
Default is NULL, in which case weighting will be done by channel means.
Only used for WLS}

\item{output.dir}{Directory to save the unmixed FCS files (default is asp$unmixed.fcs.dir).}

\item{file.suffix}{A character string to append to the output file name.
Default is NULL.}

\item{include.raw}{Logical indicating whether to include raw data in the output.
Default is FALSE.}

\item{include.imaging}{Logical indicating whether to include imaging data in
the output (relevant for S8 and A8). Default is FALSE.}

\item{divergence.threshold}{Numeric. Used for FastPoisson only. Threshold to
trigger reversion towards WLS unmixing when Poisson result diverge.}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are "NonNeg" (non-negativity will be enforced), "WLS" (revert
to WLS intial unmix) or "Balance" (WLS and NonNeg will be averaged).
Default is "Balance".}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells. Used
for "Balance" option under divergence.handling. Default is 0.5.}
}
\value{
None. Saves the unmixed FCS files to the specified output directory.
}
\description{
This function unmixes all FCS files in a specified directory
    using the provided spectra and method, and saves the unmixed data to an
    output directory.
}
