% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_folder.R
\name{unmix.folder}
\alias{unmix.folder}
\title{Unmix All FCS Files in a Directory}
\usage{
unmix.folder(
  fcs.dir,
  spectra,
  asp,
  flow.control,
  method = "OLS",
  weights = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5
)
}
\arguments{
\item{fcs.dir}{Directory containing FCS files to be unmixed.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
Options are \code{OLS}, \code{WLS}, \code{Poisson} and \code{FastPoisson}. Default is \code{OLS}.
\code{FastPoisson} requires installation of \code{AutoSpectralRcpp}}

\item{weights}{Optional numeric vector of weights: one per fluorescent
detector. Default is \code{NULL}, in which case weighting will be done by
channel means. Only used for \code{WLS}}

\item{output.dir}{Directory to save the unmixed FCS files
(default is asp$unmixed.fcs.dir).}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}}

\item{include.raw}{Logical indicating whether to include raw data in the
written FCS file. Default is \code{FALSE}}

\item{include.imaging}{Logical indicating whether to include imaging data in
the written FCS file: relevant for S8 and A8. Default is \code{FALSE}}

\item{divergence.threshold}{Numeric. Used for \code{FastPoisson} only. Threshold
to trigger reversion towards WLS unmixing when Poisson result diverge.
Default is \code{1e4}}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are \code{NonNeg}, in which case non-negativity will be enforced,
\code{WLS}, where values will revert to the WLS initial unmix or \code{Balance},
where \code{WLS} and \code{NonNeg} results will be averaged. Default is \code{Balance}}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells.
Used for \code{Balance} option under \code{divergence.handling}. Default is \code{0.5}}
}
\value{
None. Saves the unmixed FCS files to the specified output directory.
}
\description{
This function unmixes all FCS files in a specified directory
using the provided spectra and method, and saves the unmixed data to an
output directory.
}
