% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_folder.R
\name{unmix.folder}
\alias{unmix.folder}
\title{Unmix All FCS Files in a Directory}
\usage{
unmix.folder(
  fcs.dir,
  spectra,
  asp,
  flow.control,
  method = "Automatic",
  weighted = FALSE,
  weights = NULL,
  af.spectra = NULL,
  spectra.variants = NULL,
  output.dir = NULL,
  file.suffix = NULL,
  include.raw = FALSE,
  include.imaging = FALSE,
  calculate.error = FALSE,
  use.dist0 = TRUE,
  divergence.threshold = 10000,
  divergence.handling = "Balance",
  balance.weight = 0.5,
  speed = "fast",
  parallel = FALSE,
  threads = NULL
)
}
\arguments{
\item{fcs.dir}{Directory containing FCS files to be unmixed.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{method}{A character string specifying the unmixing method to use.
The default is \code{Automatic}, which uses \code{AutoSpectral} for AF extraction if
af.spectra are provided and automatically selects \code{OLS} or \code{WLS} depending
on which is normal for the given cytometer in \code{asp$cytometer}. This means
that files from the ID7000, A8 and S8 will be unmixed using \code{WLS} while
others will be unmixed with \code{OLS}. Any option can be set manually.
Manual options are \code{OLS}, \code{WLS}, \code{AutoSpectral}, \code{Poisson} and \code{FastPoisson}.
Default is \code{OLS}. \code{FastPoisson} requires installation of \code{AutoSpectralRcpp}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm in AutoSpectral unmixing.
Default is \code{FALSE} and will use OLS.}

\item{weights}{Optional numeric vector of weights: one per fluorescent
detector. Default is \code{NULL}, in which case weighting will be done by
channel means. Only used for \code{WLS}}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}. Required for \code{AutoSpectral} unmixing. Default is
\code{NULL} and will thus provoke failure if no spectra are provided and
\code{AutoSpectral} is selected.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}. Default is \code{NULL}. Used for
AutoSpectral unmixing. Required for per-cell fluorophore optimization.}

\item{output.dir}{Directory to save the unmixed FCS files
(default is asp$unmixed.fcs.dir).}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}}

\item{include.raw}{Logical indicating whether to include raw data in the
written FCS file. Default is \code{FALSE}}

\item{include.imaging}{Logical indicating whether to include imaging data in
the written FCS file: relevant for S8 and A8. Default is \code{FALSE}}

\item{calculate.error}{Logical, whether to calculate the RMSE unmixing model
accuracy and include it as a keyword in the FCS file.}

\item{use.dist0}{Logical, controls whether the selection of the optimal AF
signature for each cell is determined by which unmixing brings the fluorophore
signals closest to 0 (\code{use.dist0} = \code{TRUE}) or by which unmixing minimizes the
per-cell residual (\code{use.dist0} = \code{FALSE}). Default is \code{TRUE}. Used for
AutoSpectral unmixing.}

\item{divergence.threshold}{Numeric. Used for \code{FastPoisson} only. Threshold
to trigger reversion towards WLS unmixing when Poisson result diverges.
Default is \code{1e4}}

\item{divergence.handling}{String. How to handle divergent cells from Poisson
IRLS. Options are \code{NonNeg}, in which case non-negativity will be enforced,
\code{WLS}, where values will revert to the WLS initial unmix or \code{Balance},
where \code{WLS} and \code{NonNeg} results will be averaged. Default is \code{Balance}}

\item{balance.weight}{Numeric. Weighting to average non-convergent cells.
Used for \code{Balance} option under \code{divergence.handling}. Default is \code{0.5}}

\item{speed}{Selector for the precision-speed trade-off in AutoSpectral per-cell
fluorophore optimization. Options are the default \code{fast}, which selects the
best spectral fit per cell by updating the predicted values for each
fluorophore independently without repeating the unnmixing, \code{medium} which uses
a Woodbury-Sherman-Morrison rank-one updating of the unnmixing matrix for
better results and a moderate slow-down, or \code{slow}, which explicitly
recomputes the unmixing matrix for each variant for maximum precision. The
\code{fast} method is only one available in the \code{AutoSpectral} package and will be
slow in the pure R implementation. Installation of \code{AutoSpectralRcpp} is
strongly encouraged.}

\item{parallel}{Logical, default is \code{FALSE}. Set to \code{TRUE} to activate parallel
processing for multiple FCS files.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}.}
}
\value{
None. Saves the unmixed FCS files to the specified output directory.
}
\description{
This function unmixes all FCS files in a specified directory using the
provided spectra and method, and saves the unmixed FCS files to an output
directory of the user's choice.
}
