% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_poisson.R
\name{unmix.poisson}
\alias{unmix.poisson}
\title{Unmix Using Poisson Regression}
\usage{
unmix.poisson(
  raw.data,
  spectra,
  asp,
  initial.weights = NULL,
  parallel = TRUE,
  threads = NULL
)
}
\arguments{
\item{raw.data}{Matrix containing raw data to be unmixed.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{initial.weights}{Optional numeric vector of weights, one per fluorescent
detector. Default is \code{NULL}, in which case weighting will be done by
channel means.}

\item{parallel}{Logical, default is \code{TRUE}, which enables parallel processing
for per-cell unmixing.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used if \code{parallel=TRUE}.}
}
\value{
A matrix containing the unmixed data.
}
\description{
This function performs unmixing of raw data using Poisson regression, with
iterative reweighted least squares (IRLS) and fallback methods for cells
that fail to converge.
}
\references{
Novo D. et al. (2014). "Generalized Unmixing Model for Multispectral Flow
Cytometry Utilizing Nonsquare Compensation Matrices" \emph{Cytometry Part A},
83(5):508–520. \doi{10.1002/cyto.a.22272}
}
