% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_wls_per_cell.R
\name{unmix.wls.per.cell}
\alias{unmix.wls.per.cell}
\title{Unmix WLS Per Cell}
\usage{
unmix.wls.per.cell(cell.raw, spectra, weights)
}
\arguments{
\item{cell.raw}{Expression data from raw FCS files. A single row of cells with
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{weights}{The weighting values for the detectors. No checks are performed
in this function; \code{unmix.wls} should be used for standard cases.}
}
\value{
Unmixed data (one row), fluorophores in columns.
}
\description{
Faster solver for per-cell optimization workflow. Performs spectral unmixing
using weighted least squares.
}
