% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_control_file.R
\name{validate.control.file}
\alias{validate.control.file}
\title{Validate Control File}
\usage{
validate.control.file(
  control.dir,
  control.def.file,
  asp,
  min.event.warning,
  min.event.error,
  legacy = FALSE
)
}
\arguments{
\item{control.dir}{File path to the single-stained control FCS files.}

\item{control.def.file}{CSV file defining the single-color control file names,
fluorophores they represent, marker names, peak channels, and gating requirements.}

\item{asp}{The AutoSpectral parameter list.}

\item{min.event.warning}{The number of events in the entire FCS file that will
trigger a warning if not met.}

\item{min.event.error}{The number of events in the entire FCS file that will
trigger an error if not met.}

\item{legacy}{Logical. If \code{FALSE}, gating-related columns will not be created
and the control file will be suitable only for the new automated spectral
extraction pipeline using \code{get.spectra.automated()}. To use the version 1
"legacy" pipeline for the extraction of fluorophore spectra, using gating and
\code{define.flow.control()}, set \code{legacy=TRUE}.}
}
\value{
A dataframe of errors and warnings intended to help the user fix
problems with the \code{control.def.file}.
}
\description{
Checks whether the control file used to define the single-stained control
setup for AutoSpectral conforms to expectations and follows rules required for
successful running of AutoSpectral.
}
