% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_control_file.R
\name{validate.control.file}
\alias{validate.control.file}
\title{Validate Control File}
\usage{
validate.control.file(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{File path to the single-stained control FCS files.}

\item{control.def.file}{CSV file defining the single-color control file names,
fluorophores they represent, marker names, peak channels, and gating requirements.}

\item{asp}{The AutoSpectral parameter list.}
}
\value{
A dataframe of errors and warnings intended to help the user fix
problems with the \code{control.def.file}.
}
\description{
Checks whether the control file used to define the single-stained control
setup for AutoSpectral conforms to expectations and follows rules required for
successful running of AutoSpectral.
}
